refactor(stmt2): modify stmt2 api
