+++ taos-ws/src/stmt2/bind.rs
@@ -1,1780 +1,2955 @@
 use byteorder::{ByteOrder, LittleEndian};
-use bytes::{BufMut, BytesMut};
 use taos_query::{
-    common::{BorrowedValue, ColumnView},
+    common::{BorrowedValue, ColumnView, Value},
     stmt2::Stmt2BindData,
     RawResult,
 };
 use tracing::debug;
 
-const TOTAL_LENGTH_POSITION: usize = 0;
-const TABLE_COUNT_POSITION: usize = TOTAL_LENGTH_POSITION + 4;
-const TAG_COUNT_POSITION: usize = TABLE_COUNT_POSITION + 4;
-const COL_COUNT_POSITION: usize = TAG_COUNT_POSITION + 4;
-const TABLE_NAMES_OFFSET_POSITION: usize = COL_COUNT_POSITION + 4;
-const TAGS_OFFSET_POSITION: usize = TABLE_NAMES_OFFSET_POSITION + 4;
-const COLS_OFFSET_POSITION: usize = TAGS_OFFSET_POSITION + 4;
-const DATA_POSITION: usize = COLS_OFFSET_POSITION + 4;
+use super::{BindType, PrepareField, ReqId, StmtId};
 
-const BIND_DATA_TOTAL_LENGTH_OFFSET: usize = 0;
-const BIND_DATA_TYPE_OFFSET: usize = BIND_DATA_TOTAL_LENGTH_OFFSET + 4;
-const BIND_DATA_NUM_OFFSET: usize = BIND_DATA_TYPE_OFFSET + 4;
-const BIND_DATA_IS_NULL_OFFSET: usize = BIND_DATA_NUM_OFFSET + 4;
+const REQ_ID_POS: usize = 0;
+const STMT_ID_POS: usize = REQ_ID_POS + 8;
+const ACTION_POS: usize = STMT_ID_POS + 8;
+const VERSION_POS: usize = ACTION_POS + 8;
+const COL_IDX_POS: usize = VERSION_POS + 2;
+const FIXED_HEADER_LEN: usize = COL_IDX_POS + 4;
 
-pub(crate) fn bind_datas_as_bytes(datas: &[Stmt2BindData], is_insert: bool) -> RawResult<Vec<u8>> {
-    let table_cnt = datas.len();
-    if table_cnt == 0 {
-        return Err("Empty data".into());
+const TOTAL_LENGTH_POS: usize = 0;
+const TABLE_COUNT_POS: usize = TOTAL_LENGTH_POS + 4;
+const TAG_COUNT_POS: usize = TABLE_COUNT_POS + 4;
+const COL_COUNT_POS: usize = TAG_COUNT_POS + 4;
+const TABLE_NAMES_OFFSET_POS: usize = COL_COUNT_POS + 4;
+const TAGS_OFFSET_POS: usize = TABLE_NAMES_OFFSET_POS + 4;
+const COLS_OFFSET_POS: usize = TAGS_OFFSET_POS + 4;
+const DATA_POS: usize = COLS_OFFSET_POS + 4;
+
+// `TC` (Tag and Column) is an abbreviation used to represent tag and column.
+const TC_DATA_TOTAL_LENGTH_POS: usize = 0;
+const TC_DATA_TYPE_POS: usize = TC_DATA_TOTAL_LENGTH_POS + 4;
+const TC_DATA_NUM_POS: usize = TC_DATA_TYPE_POS + 4;
+const TC_DATA_IS_NULL_POS: usize = TC_DATA_NUM_POS + 4;
+
+const ACTION: u64 = 9;
+const VERSION: u16 = 1;
+const COL_IDX: i32 = -1;
+
+pub(super) fn bind_datas_as_bytes(
+    datas: &[Stmt2BindData],
+    req_id: ReqId,
+    stmt_id: StmtId,
+    is_insert: bool,
+    fields: Option<&Vec<PrepareField>>,
+    fields_count: usize,
+) -> RawResult<Vec<u8>> {
+    if datas.is_empty() {
+        return Err("empty data".into());
     }
 
-    let mut have_table_names = false;
+    let mut need_tbnames = false;
+    let mut need_tags = false;
+    let mut need_cols = false;
 
+    let table_cnt = datas.len();
     let mut tag_cnt = 0;
     let mut col_cnt = 0;
 
     if is_insert {
-        for data in datas {
-            if data.table_name().map_or(false, |s| !s.is_empty()) {
-                have_table_names = true;
-            }
-
-            if let Some(views) = data.tags() {
-                tag_cnt = views.len();
-                if views.iter().any(|view| view.len() != 1) {
-                    return Err("Tag can only have one row".into());
-                };
-            }
-
-            let columns = data.columns();
-            col_cnt = columns.len();
-            for i in 1..col_cnt {
-                if columns[i - 1].len() != columns[i].len() {
-                    return Err("Columns row count not match".into());
+        if let Some(fields) = fields {
+            for field in fields {
+                match field.bind_type {
+                    BindType::TableName => {
+                        need_tbnames = true;
+                    }
+                    BindType::Tag => {
+                        need_tags = true;
+                        tag_cnt += 1;
+                    }
+                    BindType::Column => {
+                        need_cols = true;
+                        col_cnt += 1;
+                    }
                 }
             }
-
-            // TODO: check type use stmt2_get_fields
         }
     } else {
-        if table_cnt != 1 {
-            return Err("Query can only has one table".into());
-        }
-
-        let data = &datas[0];
-        if data.table_name().is_some() {
-            return Err("Query not needs table name".into());
-        }
-        if data.tags().is_some() {
-            return Err("Query not needs tags".into());
-        }
-
-        let columns = data.columns();
-        if columns.len() == 0 {
-            return Err("Query needs columns".into());
-        }
-        for (i, view) in columns.iter().enumerate() {
-            if view.len() != 1 {
-                return Err(format!(
-                    "Query columns must be one row, column: {}, count: {}",
-                    i,
-                    view.len()
-                )
-                .as_str()
-                .into());
-            }
-            if view.get(0).unwrap().is_null() {
-                return Err("Query column cannot be null".into());
-            }
-        }
-
-        col_cnt = columns.len();
+        need_cols = true;
+        col_cnt = fields_count;
     }
 
-    let have_tags = tag_cnt != 0;
-    let have_cols = col_cnt != 0;
-
+    debug!("need_tbnames: {need_tbnames}, need_tags: {need_tags}, need_cols: {need_cols}");
     debug!("table_cnt: {table_cnt}, tag_cnt: {tag_cnt}, col_cnt: {col_cnt}");
-    debug!("have_table_names: {have_table_names}, have_tags: {have_tags}, have_cols: {have_cols}");
 
-    if !have_table_names && !have_tags && !have_cols {
-        return Err("No data".into());
+    let mut tbname_lens = vec![];
+    let mut tbname_buf_len = 0;
+    if need_tbnames {
+        tbname_lens = get_tbname_lens(datas)?;
+        tbname_buf_len = tbname_lens.iter().map(|&x| x as usize).sum();
     }
 
-    let mut tmp_buf = BytesMut::new();
-
-    let mut table_name_buf = BytesMut::new();
-    let mut table_name_lens = if have_table_names {
-        vec![0; table_cnt]
-    } else {
-        vec![]
-    };
-
-    let mut tag_buf = BytesMut::new();
-    let mut tag_data_lens = if have_tags {
-        vec![0; table_cnt]
-    } else {
-        vec![]
-    };
-
-    let mut col_buf = BytesMut::new();
-    let mut col_data_lens = if have_cols {
-        vec![0; table_cnt]
-    } else {
-        vec![]
-    };
-
-    for (i, data) in datas.iter().enumerate() {
-        if have_table_names {
-            let mut len = 0;
-            if let Some(tbname) = data.table_name() {
-                if !tbname.is_empty() {
-                    if tbname.len() > (u16::MAX - 1) as _ {
-                        return Err(format!(
-                            "Table name too long, index: {}, len: {}",
-                            i,
-                            tbname.len()
-                        )
-                        .as_str()
-                        .into());
-                    }
-
-                    len = tbname.len();
-                    table_name_buf.extend_from_slice(tbname.as_bytes());
-                }
-            }
-
-            table_name_buf.put_u8(0);
-            table_name_lens[i] = len + 1;
-        }
-
-        if have_tags {
-            let mut len = 0;
-            for tag in data.tags().unwrap() {
-                let tag_data_buf = generate_bind_insert_data(tag, &mut tmp_buf)?;
-                len += tag_data_buf.len();
-                tag_buf.extend_from_slice(&tag_data_buf);
-            }
-            tag_data_lens[i] = len;
-        }
-
-        if have_cols {
-            let mut len = 0;
-            for col in data.columns() {
-                let col_data_buf = if is_insert {
-                    generate_bind_insert_data(col, &mut tmp_buf)?
-                } else {
-                    generate_bind_query_data(col)?
-                };
-                len += col_data_buf.len();
-                col_buf.extend_from_slice(&col_data_buf);
-            }
-            col_data_lens[i] = len;
-        }
+    let mut tag_lens = vec![];
+    let mut tag_buf_len = 0;
+    if need_tags {
+        tag_lens = get_tag_lens(datas, tag_cnt)?;
+        tag_buf_len = tag_lens.iter().map(|&x| x as usize).sum();
     }
 
-    let tags_offset = DATA_POSITION + table_name_buf.len() + table_name_lens.len() * 2;
-    let cols_offset = tags_offset + tag_buf.len() + tag_data_lens.len() * 4;
-    let total_len = cols_offset + col_buf.len() + col_data_lens.len() * 4;
+    let mut col_lens = vec![];
+    let mut col_buf_len = 0;
+    if need_cols {
+        col_lens = get_col_lens(datas, col_cnt)?;
+        col_buf_len = col_lens.iter().map(|&x| x as usize).sum();
+    }
 
-    let mut buf = vec![0u8; total_len];
+    let tbname_total_len = tbname_lens.len() * 2 + tbname_buf_len;
+    let tag_total_len = tag_lens.len() * 4 + tag_buf_len;
+    let col_total_len = col_lens.len() * 4 + col_buf_len;
+    let total_len = DATA_POS + tbname_total_len + tag_total_len + col_total_len;
 
+    debug!("tbname_total_len: {tbname_total_len}, tbname_buf_len: {tbname_buf_len}");
+    debug!("tag_total_len: {tag_total_len}, tag_buf_len: {tag_buf_len}");
+    debug!("col_total_len: {col_total_len}, col_buf_len: {col_buf_len}");
     debug!("total_len: {total_len}");
-    debug!(
-        "table_name_buf_len: {}, tag_buf_len: {}, col_buf_len: {}",
-        table_name_buf.len(),
-        tag_buf.len(),
-        col_buf.len()
-    );
-    debug!(
-        "table_name_lens_len: {}, tag_data_lens_len: {}, col_data_lens_len: {}",
-        table_name_lens.len(),
-        tag_data_lens.len(),
-        col_data_lens.len()
-    );
-    debug!("tags_offset: {tags_offset}, cols_offset: {cols_offset}");
 
-    // write total length
-    LittleEndian::write_u32(&mut buf, total_len as _);
+    let mut data = vec![0u8; FIXED_HEADER_LEN + total_len];
+    write_fixed_headers(&mut data, req_id, stmt_id);
 
-    // write table count
-    LittleEndian::write_u32(&mut buf[TABLE_COUNT_POSITION..], table_cnt as _);
+    let bytes = &mut data[FIXED_HEADER_LEN..];
+    LittleEndian::write_u32(&mut bytes[TOTAL_LENGTH_POS..], total_len as _);
+    LittleEndian::write_i32(&mut bytes[TABLE_COUNT_POS..], table_cnt as _);
+    LittleEndian::write_i32(&mut bytes[TAG_COUNT_POS..], tag_cnt as _);
+    LittleEndian::write_i32(&mut bytes[COL_COUNT_POS..], col_cnt as _);
 
-    if have_table_names {
-        // write table names offset
-        LittleEndian::write_u32(&mut buf[TABLE_NAMES_OFFSET_POSITION..], DATA_POSITION as _);
-        let mut offset = DATA_POSITION;
-        // write table name length
-        for len in table_name_lens {
-            LittleEndian::write_u16(&mut buf[offset..], len as _);
-            offset += 2;
-        }
-        // write table name buffer
-        buf[offset..offset + table_name_buf.len()].copy_from_slice(&table_name_buf);
+    if need_tbnames {
+        LittleEndian::write_u32(&mut bytes[TABLE_NAMES_OFFSET_POS..], DATA_POS as _);
+        write_tbnames(&mut bytes[DATA_POS..], datas, &tbname_lens);
     }
 
-    if have_tags {
-        // write tag count
-        LittleEndian::write_u32(&mut buf[TAG_COUNT_POSITION..], tag_cnt as _);
-        // write tags offset
-        LittleEndian::write_u32(&mut buf[TAGS_OFFSET_POSITION..], tags_offset as _);
-        let mut offset = tags_offset;
-        // write tags data length
-        for len in tag_data_lens {
-            LittleEndian::write_u32(&mut buf[offset..], len as _);
-            offset += 4;
-        }
-        // write tags buffer
-        buf[offset..offset + tag_buf.len()].copy_from_slice(&tag_buf);
+    if need_tags {
+        let tags_offset = DATA_POS + tbname_total_len;
+        LittleEndian::write_u32(&mut bytes[TAGS_OFFSET_POS..], tags_offset as _);
+        write_tags(&mut bytes[tags_offset..], datas, &tag_lens);
     }
 
-    if have_cols {
-        // write col count
-        LittleEndian::write_u32(&mut buf[COL_COUNT_POSITION..], col_cnt as _);
-        // write cols offset
-        LittleEndian::write_u32(&mut buf[COLS_OFFSET_POSITION..], cols_offset as _);
-        let mut offset = cols_offset;
-        // write col data length
-        for len in col_data_lens {
-            LittleEndian::write_u32(&mut buf[offset..], len as _);
-            offset += 4;
-        }
-        // write col buffer
-        buf[offset..offset + col_buf.len()].copy_from_slice(&col_buf);
+    if need_cols {
+        let cols_offset = DATA_POS + tbname_total_len + tag_total_len;
+        LittleEndian::write_u32(&mut bytes[COLS_OFFSET_POS..], cols_offset as _);
+        write_cols(&mut bytes[cols_offset..], datas, &col_lens);
     }
 
-    Ok(buf)
+    Ok(data)
 }
 
-fn generate_bind_insert_data(view: &ColumnView, tmp_buf: &mut BytesMut) -> RawResult<Vec<u8>> {
-    tmp_buf.clear();
-
-    let num = view.len();
-    let ty = view.as_ty();
-    let have_len = ty.fixed_length() == 0;
-    let header_len = get_bind_data_header_len(num, have_len);
-
-    let mut is_nulls = vec![0u8; num];
-    let mut lens = vec![0u8; num * 4];
-
-    if check_all_null(view) {
-        is_nulls.iter_mut().for_each(|x| *x = 1);
-    } else {
-        use ColumnView::*;
-        match view {
-            Bool(bool_view) => {
-                for (i, v) in bool_view.iter().enumerate() {
-                    match v {
-                        Some(b) => {
-                            if b {
-                                tmp_buf.put_u8(1);
-                            } else {
-                                tmp_buf.put_u8(0);
-                            }
-                        }
-                        None => {
-                            is_nulls[i] = 1;
-                            tmp_buf.put_u8(0);
-                        }
-                    }
-                }
-            }
-            TinyInt(tinyint_view) => {
-                for (i, v) in tinyint_view.iter().enumerate() {
-                    if v.is_none() {
-                        is_nulls[i] = 1;
-                    }
-                    tmp_buf.put_i8(v.unwrap_or(0));
-                }
-            }
-            SmallInt(smallint_view) => {
-                for (i, v) in smallint_view.iter().enumerate() {
-                    if v.is_none() {
-                        is_nulls[i] = 1;
-                    }
-                    tmp_buf.put_i16_le(v.unwrap_or(0));
-                }
-            }
-            Int(int_view) => {
-                for (i, v) in int_view.iter().enumerate() {
-                    if v.is_none() {
-                        is_nulls[i] = 1;
-                    }
-                    tmp_buf.put_i32_le(v.unwrap_or(0));
-                }
-            }
-            BigInt(big_int_view) => {
-                for (i, v) in big_int_view.iter().enumerate() {
-                    if v.is_none() {
-                        is_nulls[i] = 1;
-                    }
-                    tmp_buf.put_i64_le(v.unwrap_or(0));
-                }
-            }
-            Float(float_view) => {
-                for (i, v) in float_view.iter().enumerate() {
-                    if v.is_none() {
-                        is_nulls[i] = 1;
-                    }
-                    tmp_buf.put_f32_le(v.unwrap_or(0f32));
-                }
-            }
-            Double(double_view) => {
-                for (i, v) in double_view.iter().enumerate() {
-                    if v.is_none() {
-                        is_nulls[i] = 1;
-                    }
-                    tmp_buf.put_f64_le(v.unwrap_or(0f64));
-                }
-            }
-            VarChar(varchar_view) => {
-                for (i, v) in varchar_view.iter().enumerate() {
-                    let mut len = 0;
-                    match v {
-                        Some(is) => {
-                            tmp_buf.extend_from_slice(is.as_bytes());
-                            len = is.len();
-                        }
-                        None => is_nulls[i] = 1,
-                    };
-                    LittleEndian::write_u32(&mut lens[i * 4..], len as _);
-                }
-            }
-            Timestamp(timestamp_view) => {
-                for (i, v) in timestamp_view.iter().enumerate() {
-                    if v.is_none() {
-                        is_nulls[i] = 1;
-                    }
-                    tmp_buf.put_i64_le(
-                        v.unwrap_or(taos_query::common::Timestamp::Microseconds(0))
-                            .as_raw_i64(),
-                    );
-                }
-            }
-            NChar(nchar_view) => {
-                for (i, v) in nchar_view.iter().enumerate() {
-                    let mut len = 0;
-                    match v {
-                        Some(is) => {
-                            tmp_buf.extend_from_slice(is.as_bytes());
-                            len = is.len();
-                        }
-                        None => is_nulls[i] = 1,
-                    }
-                    LittleEndian::write_u32(&mut lens[i * 4..], len as _);
-                }
-            }
-            UTinyInt(utinyint_view) => {
-                for (i, v) in utinyint_view.iter().enumerate() {
-                    if v.is_none() {
-                        is_nulls[i] = 1;
-                    }
-                    tmp_buf.put_u8(v.unwrap_or(0));
-                }
-            }
-            USmallInt(usmallint_view) => {
-                for (i, v) in usmallint_view.iter().enumerate() {
-                    if v.is_none() {
-                        is_nulls[i] = 1;
-                    }
-                    tmp_buf.put_u16_le(v.unwrap_or(0));
-                }
-            }
-            UInt(uint_view) => {
-                for (i, v) in uint_view.iter().enumerate() {
-                    if v.is_none() {
-                        is_nulls[i] = 1;
-                    }
-                    tmp_buf.put_u32_le(v.unwrap_or(0));
-                }
-            }
-            UBigInt(ubigint_view) => {
-                for (i, v) in ubigint_view.iter().enumerate() {
-                    if v.is_none() {
-                        is_nulls[i] = 1;
-                    }
-                    tmp_buf.put_u64_le(v.unwrap_or(0));
-                }
-            }
-            Json(json_view) => {
-                for (i, v) in json_view.iter().enumerate() {
-                    let mut len = 0;
-                    match v {
-                        Some(is) => {
-                            tmp_buf.extend_from_slice(is.as_bytes());
-                            len = is.len();
-                        }
-                        None => is_nulls[i] = 1,
-                    }
-                    LittleEndian::write_u32(&mut lens[i * 4..], len as _);
-                }
-            }
-            VarBinary(varbinary_view) => {
-                for (i, v) in varbinary_view.iter().enumerate() {
-                    let mut len = 0;
-                    match v {
-                        Some(is) => {
-                            tmp_buf.extend_from_slice(is.as_bytes());
-                            len = is.len();
-                        }
-                        None => is_nulls[i] = 1,
-                    }
-                    LittleEndian::write_u32(&mut lens[i * 4..], len as _);
-                }
-            }
-            Geometry(geometry_view) => {
-                for (i, v) in geometry_view.iter().enumerate() {
-                    let mut len = 0;
-                    match v {
-                        Some(is) => {
-                            tmp_buf.extend_from_slice(is.as_bytes());
-                            len = is.len();
-                        }
-                        None => is_nulls[i] = 1,
-                    }
-                    LittleEndian::write_u32(&mut lens[i * 4..], len as _);
-                }
-            }
+fn get_tbname_lens(datas: &[Stmt2BindData]) -> RawResult<Vec<u16>> {
+    let mut tbname_lens = vec![0u16; datas.len()];
+    for (i, data) in datas.iter().enumerate() {
+        if data.table_name().map_or(true, |s| s.is_empty()) {
+            return Err("table name is empty".into());
         }
+        let tbname = data.table_name().unwrap();
+        // Add 1 because the table name ends with '\0'
+        tbname_lens[i] = (tbname.len() + 1) as _;
     }
-
-    let total_len = header_len + tmp_buf.len();
-    let mut data_buf = vec![0u8; total_len];
-    // write total length
-    LittleEndian::write_u32(&mut data_buf, total_len as _);
-    // write type
-    LittleEndian::write_u32(&mut data_buf[BIND_DATA_TYPE_OFFSET..], ty as _);
-    // write num
-    LittleEndian::write_u32(&mut data_buf[BIND_DATA_NUM_OFFSET..], num as _);
-    // write isnull
-    data_buf[BIND_DATA_IS_NULL_OFFSET..BIND_DATA_IS_NULL_OFFSET + is_nulls.len()]
-        .copy_from_slice(&is_nulls);
-    if have_len {
-        // write have length
-        data_buf[BIND_DATA_IS_NULL_OFFSET + num] = 1;
-
-        // write length
-        let len_offset = BIND_DATA_IS_NULL_OFFSET + num + 1;
-        data_buf[len_offset..len_offset + lens.len()].copy_from_slice(&lens);
-    }
-    // write buffer length
-    LittleEndian::write_u32(&mut data_buf[header_len - 4..], tmp_buf.len() as _);
-    // write buffer
-    data_buf[header_len..].copy_from_slice(&tmp_buf);
-
-    Ok(data_buf)
+    Ok(tbname_lens)
 }
 
-fn generate_bind_query_data(view: &ColumnView) -> RawResult<Vec<u8>> {
-    let bval = view.get(0).unwrap();
-    let ty = bval.ty();
-    let have_len = ty.fixed_length() == 0;
-    let mut len = 0;
-    let mut buf;
+fn get_tag_lens(datas: &[Stmt2BindData], tag_cnt: usize) -> RawResult<Vec<u32>> {
+    let mut tag_lens = vec![0u32; datas.len()];
+    for (i, data) in datas.iter().enumerate() {
+        if data.tags().is_none() {
+            return Err("tags are empty".into());
+        }
 
-    use BorrowedValue::*;
-    match bval {
-        Bool(v) => {
-            buf = vec![0u8; 1];
-            buf[0] = if v { 1 } else { 0 };
+        let tags = data.tags().unwrap();
+        if tags.len() != tag_cnt {
+            return Err("tags len mismatch".into());
         }
-        TinyInt(v) => {
-            buf = vec![v as u8];
-        }
-        SmallInt(v) => {
-            buf = vec![0u8; 2];
-            LittleEndian::write_i16(&mut buf, v);
-        }
-        Int(v) => {
-            buf = vec![0u8; 4];
-            LittleEndian::write_i32(&mut buf, v);
-        }
-        BigInt(v) => {
-            buf = vec![0u8; 8];
-            LittleEndian::write_i64(&mut buf, v);
-        }
-        Float(v) => {
-            buf = vec![0u8; 4];
-            LittleEndian::write_f32(&mut buf, v);
-        }
-        Double(v) => {
-            buf = vec![0u8; 8];
-            LittleEndian::write_f64(&mut buf, v);
-        }
-        VarChar(v) => {
-            len = v.len();
-            buf = vec![0u8; len];
-            buf.copy_from_slice(v.as_bytes());
-        }
-        Timestamp(ts) => {
-            buf = vec![0u8; 8];
-            LittleEndian::write_i64(&mut buf, ts.as_raw_i64());
-        }
-        NChar(v) => {
-            len = v.len();
-            buf = vec![0u8; len];
-            buf.copy_from_slice(v.as_bytes());
-        }
-        UTinyInt(v) => {
-            buf = vec![v];
-        }
-        USmallInt(v) => {
-            buf = vec![0u8; 2];
-            LittleEndian::write_u16(&mut buf, v);
-        }
-        UInt(v) => {
-            buf = vec![0u8; 4];
-            LittleEndian::write_u32(&mut buf, v);
-        }
-        UBigInt(v) => {
-            buf = vec![0u8; 8];
-            LittleEndian::write_u64(&mut buf, v);
-        }
-        Json(v) | VarBinary(v) | Geometry(v) => {
-            len = v.len();
-            buf = vec![0u8; len];
-            buf.copy_from_slice(v.as_ref());
-        }
-        _ => {
-            return Err("unsupported type".into());
-        }
-    };
 
-    let header_len = get_bind_data_header_len(1, have_len);
-    let total_len = header_len + buf.len();
-    let mut data_buf = vec![0u8; total_len];
-    // write total length
-    LittleEndian::write_u32(&mut data_buf, total_len as u32);
-    // write type
-    LittleEndian::write_u32(&mut data_buf[BIND_DATA_TYPE_OFFSET..], ty as u32);
-    // write num
-    LittleEndian::write_u32(&mut data_buf[BIND_DATA_NUM_OFFSET..], 1);
-    // write isnull
-    data_buf[BIND_DATA_IS_NULL_OFFSET] = 0;
-    if have_len {
-        // write have length
-        data_buf[BIND_DATA_IS_NULL_OFFSET + 1] = 1;
-        // write length
-        LittleEndian::write_u32(&mut data_buf[BIND_DATA_IS_NULL_OFFSET + 2..], len as _);
+        let mut len = 0;
+        for tag in tags {
+            let have_len = tag.ty().fixed_length() == 0;
+            len += get_tc_header_len(1, have_len);
+            len += get_tag_data_len(tag);
+        }
+        tag_lens[i] = len as _;
     }
-    // write buffer length
-    LittleEndian::write_u32(&mut data_buf[header_len - 4..], buf.len() as u32);
-    // write buffer
-    data_buf[header_len..].copy_from_slice(&buf);
-
-    Ok(data_buf)
+    Ok(tag_lens)
 }
 
-fn get_bind_data_header_len(num: usize, have_len: bool) -> usize {
+fn get_col_lens(datas: &[Stmt2BindData], col_cnt: usize) -> RawResult<Vec<u32>> {
+    let mut col_lens = vec![0u32; datas.len()];
+    for (i, data) in datas.iter().enumerate() {
+        if data.columns().is_none() {
+            return Err("columns are empty".into());
+        }
+
+        let cols = data.columns().unwrap();
+        if cols.len() != col_cnt {
+            return Err("columns len mismatch".into());
+        }
+
+        let mut len = 0;
+        for col in cols {
+            let have_len = col.as_ty().fixed_length() == 0;
+            len += get_tc_header_len(col.len(), have_len);
+            len += get_col_data_len(col);
+        }
+        col_lens[i] = len as _;
+    }
+    Ok(col_lens)
+}
+
+fn get_tc_header_len(num: usize, have_len: bool) -> usize {
+    // TotalLength(4) + Type(4) + Num(4) + HaveLength(1) + BufferLength(4) + IsNull(num)
     let mut len = 17 + num;
     if have_len {
+        // Length(num * 4)
         len += num * 4;
     }
     len
 }
 
-fn check_all_null(view: &ColumnView) -> bool {
-    for bval in view.iter() {
-        if !bval.is_null() {
+fn get_tag_data_len(tag: &Value) -> usize {
+    use Value::*;
+    match tag {
+        Null(_) => 0,
+        VarChar(v) => v.len(),
+        Json(v) => serde_json::to_vec(v).unwrap().len(),
+        NChar(v) => v.len(),
+        Blob(v) => v.len(),
+        MediumBlob(v) => v.len(),
+        VarBinary(v) => v.len(),
+        Geometry(v) => v.len(),
+        _ => tag.ty().fixed_length(),
+    }
+}
+
+fn get_col_data_len(col: &ColumnView) -> usize {
+    if check_col_is_null(col) {
+        return 0;
+    }
+
+    let fixed_len = col.as_ty().fixed_length();
+    if fixed_len != 0 {
+        return fixed_len * col.len();
+    }
+
+    let mut len = 0;
+
+    use ColumnView::*;
+    match col {
+        VarChar(view) => {
+            for val in view.iter() {
+                if let Some(str) = val {
+                    len += str.len();
+                }
+            }
+        }
+        NChar(view) => {
+            for val in view.iter() {
+                if let Some(str) = val {
+                    len += str.len();
+                }
+            }
+        }
+        Json(view) => {
+            for val in view.iter() {
+                if let Some(json) = val {
+                    len += json.len();
+                }
+            }
+        }
+        VarBinary(view) => {
+            for val in view.iter() {
+                if let Some(bytes) = val {
+                    len += bytes.len();
+                }
+            }
+        }
+        Geometry(view) => {
+            for val in view.iter() {
+                if let Some(bytes) = val {
+                    len += bytes.len();
+                }
+            }
+        }
+        _ => unreachable!(),
+    }
+
+    len
+}
+
+fn check_col_is_null(col: &ColumnView) -> bool {
+    for val in col.iter() {
+        if !val.is_null() {
             return false;
         }
     }
     true
 }
 
+fn write_fixed_headers(bytes: &mut [u8], req_id: ReqId, stmt_id: StmtId) {
+    LittleEndian::write_u64(&mut bytes[REQ_ID_POS..], req_id);
+    LittleEndian::write_u64(&mut bytes[STMT_ID_POS..], stmt_id);
+    LittleEndian::write_u64(&mut bytes[ACTION_POS..], ACTION);
+    LittleEndian::write_u16(&mut bytes[VERSION_POS..], VERSION);
+    LittleEndian::write_i32(&mut bytes[COL_IDX_POS..], COL_IDX);
+}
+
+fn write_tbnames(bytes: &mut [u8], datas: &[Stmt2BindData], tbname_lens: &[u16]) {
+    // Write TableNameLength
+    let mut offset = 0;
+    for len in tbname_lens {
+        LittleEndian::write_u16(&mut bytes[offset..], *len);
+        offset += 2;
+    }
+
+    // Write TableNameBuffer
+    for data in datas {
+        let tbname = data.table_name().unwrap();
+        let len = tbname.len();
+        bytes[offset..offset + len].copy_from_slice(tbname.as_bytes());
+        // Add 1 because the table name end with '\0'
+        offset += len + 1;
+    }
+}
+
+fn write_tags(bytes: &mut [u8], datas: &[Stmt2BindData], tag_lens: &[u32]) {
+    // Write TagsDataLength
+    let mut offset = 0;
+    for len in tag_lens {
+        LittleEndian::write_u32(&mut bytes[offset..], *len);
+        offset += 4;
+    }
+
+    // Write TagsBuffer
+    for data in datas {
+        for tag in data.tags().unwrap() {
+            offset += write_tag(&mut bytes[offset..], tag);
+        }
+    }
+}
+
+fn write_tag(bytes: &mut [u8], tag: &Value) -> usize {
+    let ty = tag.ty();
+    let is_null = tag.is_null();
+    let have_len = ty.fixed_length() == 0;
+    let header_len = get_tc_header_len(1, have_len);
+
+    let offset = header_len;
+    let mut buf_len = 0;
+    if !is_null {
+        buf_len = ty.fixed_length();
+        let val = tag.to_borrowed_value();
+
+        // Write Buffer
+        use BorrowedValue::*;
+        match val {
+            Null(_) => unreachable!(),
+            Bool(v) => {
+                if v {
+                    bytes[offset] = 1;
+                }
+            }
+            TinyInt(v) => bytes[offset] = v as _,
+            UTinyInt(v) => bytes[offset] = v,
+            SmallInt(v) => LittleEndian::write_i16(&mut bytes[offset..], v as _),
+            USmallInt(v) => LittleEndian::write_u16(&mut bytes[offset..], v),
+            Int(v) => LittleEndian::write_i32(&mut bytes[offset..], v as _),
+            UInt(v) => LittleEndian::write_u32(&mut bytes[offset..], v),
+            BigInt(v) => LittleEndian::write_i64(&mut bytes[offset..], v as _),
+            UBigInt(v) => LittleEndian::write_u64(&mut bytes[offset..], v),
+            Float(v) => LittleEndian::write_f32(&mut bytes[offset..], v),
+            Double(v) => LittleEndian::write_f64(&mut bytes[offset..], v),
+            Timestamp(v) => LittleEndian::write_i64(&mut bytes[offset..], v.as_raw_i64() as _),
+            VarChar(v) => {
+                buf_len = v.len();
+                bytes[offset..offset + v.len()].copy_from_slice(v.as_bytes());
+            }
+            NChar(v) => {
+                buf_len = v.len();
+                bytes[offset..offset + v.len()].copy_from_slice(v.as_bytes());
+            }
+            Json(v) | VarBinary(v) | Geometry(v) => {
+                buf_len = v.len();
+                bytes[offset..offset + v.len()].copy_from_slice(&v);
+            }
+            _ => unreachable!("Unsupported Type"),
+        };
+
+        // Write BufferLength
+        LittleEndian::write_u32(&mut bytes[header_len - 4..], buf_len as _);
+    }
+
+    let total_len = header_len + buf_len;
+
+    LittleEndian::write_u32(&mut bytes[TC_DATA_TOTAL_LENGTH_POS..], total_len as _);
+    LittleEndian::write_u32(&mut bytes[TC_DATA_TYPE_POS..], ty as _);
+    LittleEndian::write_u32(&mut bytes[TC_DATA_NUM_POS..], 1);
+
+    if is_null {
+        bytes[TC_DATA_IS_NULL_POS] = 1;
+    }
+
+    if have_len {
+        // Write HaveLength
+        bytes[TC_DATA_IS_NULL_POS + 1] = 1;
+        // Write Length
+        LittleEndian::write_u32(&mut bytes[TC_DATA_IS_NULL_POS + 2..], buf_len as _);
+    }
+
+    total_len
+}
+
+fn write_cols(bytes: &mut [u8], datas: &[Stmt2BindData], col_lens: &[u32]) {
+    // Write ColDataLength
+    let mut offset = 0;
+    for len in col_lens {
+        LittleEndian::write_u32(&mut bytes[offset..], *len);
+        offset += 4;
+    }
+
+    // Write ColBuffer
+    for data in datas {
+        let cols = data.columns().unwrap();
+        for col in cols {
+            offset += write_col(&mut bytes[offset..], col);
+        }
+    }
+}
+
+fn write_col(bytes: &mut [u8], col: &ColumnView) -> usize {
+    let num = col.len();
+    let ty = col.as_ty();
+    let is_null = check_col_is_null(col);
+    let have_len = ty.fixed_length() == 0;
+    let header_len = get_tc_header_len(num, have_len);
+    let mut buf_offset = header_len;
+
+    if is_null {
+        // Write IsNull
+        for i in TC_DATA_IS_NULL_POS..TC_DATA_IS_NULL_POS + num {
+            bytes[i] = 1;
+        }
+    } else {
+        // Write IsNull, Length and Buffer
+        let mut is_null_offset = TC_DATA_IS_NULL_POS;
+        // TC_DATA_IS_NULL_POS + IsNull(num) + HaveLength(1)
+        let mut len_offset = is_null_offset + num + 1;
+        use ColumnView::*;
+        match col {
+            Bool(view) => {
+                for val in view.iter() {
+                    match val {
+                        Some(b) => {
+                            if b {
+                                bytes[buf_offset] = 1;
+                            }
+                        }
+                        None => bytes[is_null_offset] = 1,
+                    }
+                    buf_offset += 1;
+                    is_null_offset += 1;
+                }
+            }
+            TinyInt(view) => {
+                for val in view.iter() {
+                    if val.is_none() {
+                        bytes[is_null_offset] = 1;
+                    }
+                    bytes[buf_offset] = val.unwrap_or(0) as _;
+                    buf_offset += 1;
+                    is_null_offset += 1;
+                }
+            }
+            UTinyInt(view) => {
+                for val in view.iter() {
+                    if val.is_none() {
+                        bytes[is_null_offset] = 1;
+                    }
+                    bytes[buf_offset] = val.unwrap_or(0);
+                    buf_offset += 1;
+                    is_null_offset += 1;
+                }
+            }
+            SmallInt(view) => {
+                for val in view.iter() {
+                    if val.is_none() {
+                        bytes[is_null_offset] = 1;
+                    }
+                    LittleEndian::write_i16(&mut bytes[buf_offset..], val.unwrap_or(0));
+                    buf_offset += 2;
+                    is_null_offset += 1;
+                }
+            }
+            USmallInt(view) => {
+                for val in view.iter() {
+                    if val.is_none() {
+                        bytes[is_null_offset] = 1;
+                    }
+                    LittleEndian::write_u16(&mut bytes[buf_offset..], val.unwrap_or(0));
+                    buf_offset += 2;
+                    is_null_offset += 1;
+                }
+            }
+            Int(view) => {
+                for val in view.iter() {
+                    if val.is_none() {
+                        bytes[is_null_offset] = 1;
+                    }
+                    LittleEndian::write_i32(&mut bytes[buf_offset..], val.unwrap_or(0));
+                    buf_offset += 4;
+                    is_null_offset += 1;
+                }
+            }
+            UInt(view) => {
+                for val in view.iter() {
+                    if val.is_none() {
+                        bytes[is_null_offset] = 1;
+                    }
+                    LittleEndian::write_u32(&mut bytes[buf_offset..], val.unwrap_or(0));
+                    buf_offset += 4;
+                    is_null_offset += 1;
+                }
+            }
+            BigInt(view) => {
+                for val in view.iter() {
+                    if val.is_none() {
+                        bytes[is_null_offset] = 1;
+                    }
+                    LittleEndian::write_i64(&mut bytes[buf_offset..], val.unwrap_or(0));
+                    buf_offset += 8;
+                    is_null_offset += 1;
+                }
+            }
+            UBigInt(view) => {
+                for val in view.iter() {
+                    if val.is_none() {
+                        bytes[is_null_offset] = 1;
+                    }
+                    LittleEndian::write_u64(&mut bytes[buf_offset..], val.unwrap_or(0));
+                    buf_offset += 8;
+                    is_null_offset += 1;
+                }
+            }
+            Float(view) => {
+                for val in view.iter() {
+                    if val.is_none() {
+                        bytes[is_null_offset] = 1;
+                    }
+                    LittleEndian::write_f32(&mut bytes[buf_offset..], val.unwrap_or(0f32));
+                    buf_offset += 4;
+                    is_null_offset += 1;
+                }
+            }
+            Double(view) => {
+                for val in view.iter() {
+                    if val.is_none() {
+                        bytes[is_null_offset] = 1;
+                    }
+                    LittleEndian::write_f64(&mut bytes[buf_offset..], val.unwrap_or(0f64));
+                    buf_offset += 8;
+                    is_null_offset += 1;
+                }
+            }
+            Timestamp(view) => {
+                for val in view.iter() {
+                    if val.is_none() {
+                        bytes[is_null_offset] = 1;
+                    }
+                    let ts = val
+                        .unwrap_or(taos_query::common::Timestamp::Microseconds(0))
+                        .as_raw_i64();
+                    LittleEndian::write_i64(&mut bytes[buf_offset..], ts);
+                    buf_offset += 8;
+                    is_null_offset += 1;
+                }
+            }
+            VarChar(view) => {
+                for val in view.iter() {
+                    let mut len = 0;
+                    match val {
+                        Some(str) => {
+                            len = str.len();
+                            bytes[buf_offset..buf_offset + len].copy_from_slice(str.as_bytes());
+                            buf_offset += len;
+                        }
+                        None => bytes[is_null_offset] = 1,
+                    }
+                    LittleEndian::write_i32(&mut bytes[len_offset..], len as _);
+                    is_null_offset += 1;
+                    len_offset += 4;
+                }
+            }
+            NChar(view) => {
+                for val in view.iter() {
+                    let mut len = 0;
+                    match val {
+                        Some(str) => {
+                            len = str.len();
+                            bytes[buf_offset..buf_offset + len].copy_from_slice(str.as_bytes());
+                            buf_offset += len;
+                        }
+                        None => bytes[is_null_offset] = 1,
+                    }
+                    LittleEndian::write_i32(&mut bytes[len_offset..], len as _);
+                    is_null_offset += 1;
+                    len_offset += 4;
+                }
+            }
+            Json(view) => {
+                for val in view.iter() {
+                    let mut len = 0;
+                    match val {
+                        Some(json) => {
+                            len = json.len();
+                            bytes[buf_offset..buf_offset + len].copy_from_slice(json.as_bytes());
+                            buf_offset += len;
+                        }
+                        None => bytes[is_null_offset] = 1,
+                    }
+                    LittleEndian::write_i32(&mut bytes[len_offset..], len as _);
+                    is_null_offset += 1;
+                    len_offset += 4;
+                }
+            }
+            VarBinary(view) => {
+                for val in view.iter() {
+                    let mut len = 0;
+                    match val {
+                        Some(bs) => {
+                            len = bs.len();
+                            bytes[buf_offset..buf_offset + len].copy_from_slice(bs.as_bytes());
+                            buf_offset += len;
+                        }
+                        None => bytes[is_null_offset] = 1,
+                    }
+                    LittleEndian::write_i32(&mut bytes[len_offset..], len as _);
+                    is_null_offset += 1;
+                    len_offset += 4;
+                }
+            }
+            Geometry(view) => {
+                for val in view.iter() {
+                    let mut len = 0;
+                    match val {
+                        Some(bs) => {
+                            len = bs.len();
+                            bytes[buf_offset..buf_offset + len].copy_from_slice(bs.as_bytes());
+                            buf_offset += len;
+                        }
+                        None => bytes[is_null_offset] = 1,
+                    }
+                    LittleEndian::write_i32(&mut bytes[len_offset..], len as _);
+                    is_null_offset += 1;
+                    len_offset += 4;
+                }
+            }
+        }
+    }
+
+    let total_len = buf_offset;
+
+    LittleEndian::write_u32(&mut bytes[TC_DATA_TOTAL_LENGTH_POS..], total_len as _);
+    LittleEndian::write_u32(&mut bytes[TC_DATA_TYPE_POS..], ty as _);
+    LittleEndian::write_u32(&mut bytes[TC_DATA_NUM_POS..], num as _);
+
+    if have_len {
+        // Write HaveLenght
+        bytes[TC_DATA_IS_NULL_POS + num] = 1;
+    }
+
+    if !is_null {
+        // Write BufferLength
+        let buf_len = total_len - header_len;
+        LittleEndian::write_u32(&mut bytes[header_len - 4..], buf_len as _);
+    }
+
+    total_len
+}
+
 #[cfg(test)]
 mod tests {
-    use anyhow::Result;
-    use ctor::ctor;
+    use bytes::Bytes;
+    use taos_query::common::{ColumnView, Timestamp, Ty, Value};
 
-    use super::*;
+    use crate::stmt2::{BindType, PrepareField};
 
-    #[ctor]
-    fn setup() {
-        tracing_subscriber::fmt::init();
+    use super::{bind_datas_as_bytes, Stmt2BindData};
+
+    #[test]
+    fn test_bind_datas_as_bytes_with_tbnames() -> anyhow::Result<()> {
+        let data1 = Stmt2BindData::new(Some("test1"), None, None);
+        let data2 = Stmt2BindData::new(Some("test2"), None, None);
+        let data3 = Stmt2BindData::new(Some("test3"), None, None);
+        let datas = [data1, data2, data3];
+
+        let fields = vec![PrepareField {
+            name: "".to_string(),
+            field_type: 1,
+            precision: 0,
+            scale: 0,
+            bytes: 131584,
+            bind_type: BindType::TableName,
+        }];
+
+        let res = bind_datas_as_bytes(&datas, 100, 200, true, Some(&fields), 0)?;
+
+        #[rustfmt::skip]
+        let expected = [
+            // fixed headers
+            0x64, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // req_id
+            0xc8, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // stmt_id
+            0x09, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // action
+            0x01, 0x00, // version
+            0xff, 0xff, 0xff, 0xff, // col_idx
+
+            // data
+            0x34, 0x00, 0x00, 0x00, // TotalLength
+            0x03, 0x00, 0x00, 0x00, // TableCount
+            0x00, 0x00, 0x00, 0x00, // TagCount
+            0x00, 0x00, 0x00, 0x00, // ColCount
+            0x1c, 0x00, 0x00, 0x00, // TableNamesOffset
+            0x00, 0x00, 0x00, 0x00, // TagsOffset
+            0x00, 0x00, 0x00, 0x00, // ColsOffset
+
+            // table names
+            // TableNameLength
+            0x06, 0x00,
+            0x06, 0x00,
+            0x06, 0x00,
+            // TableNameBuffer
+            0x74, 0x65, 0x73, 0x74, 0x31, 0x00,
+            0x74, 0x65, 0x73, 0x74, 0x32, 0x00,
+            0x74, 0x65, 0x73, 0x74, 0x33, 0x00,
+        ];
+
+        assert_eq!(res, expected);
+
+        Ok(())
     }
 
     #[test]
-    fn test_bind_datas_as_bytes() -> Result<()> {
-        {
-            let data1 = Stmt2BindData::new(Some("test1"), None, &[]);
-            let data2 = Stmt2BindData::new(Some(""), None, &[]);
-            let data3 = Stmt2BindData::new(Some("test2"), None, &[]);
-            let res = bind_datas_as_bytes(&[data1, data2, data3], true)?;
-            let expected = [
-                0x2f, 0x00, 0x00, 0x00, // total length
-                0x03, 0x00, 0x00, 0x00, // table count
-                0x00, 0x00, 0x00, 0x00, // tag count
-                0x00, 0x00, 0x00, 0x00, // col count
-                0x1c, 0x00, 0x00, 0x00, // table names offset
-                0x00, 0x00, 0x00, 0x00, // tags offset
-                0x00, 0x00, 0x00, 0x00, // col offset
-                // table names
-                0x06, 0x00, 0x01, 0x00, 0x06, 0x00, // table name length
-                // table name buffer
-                0x74, 0x65, 0x73, 0x74, 0x31, 0x00, // test1
-                0x00, // nil
-                0x74, 0x65, 0x73, 0x74, 0x32, 0x00, // test2
-            ];
-            assert_eq!(res, expected);
-        }
+    fn test_bind_datas_as_bytes_with_tags_1() -> anyhow::Result<()> {
+        let tags = &[
+            Value::Timestamp(Timestamp::Milliseconds(1726803356466)),
+            Value::Bool(true),
+            Value::TinyInt(1),
+            Value::SmallInt(2),
+            Value::Int(3),
+            Value::BigInt(4),
+            Value::Float(5.5),
+            Value::Double(6.6),
+            Value::UTinyInt(7),
+            Value::USmallInt(8),
+            Value::UInt(9),
+            Value::UBigInt(10),
+            Value::VarChar("binary".to_string()),
+            Value::NChar("nchar".to_string()),
+            Value::Geometry(Bytes::from(vec![
+                0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
+            ])),
+            Value::VarBinary(Bytes::from("varbinary".as_bytes())),
+        ];
 
-        {
-            let tags1 = &[
-                ColumnView::from_millis_timestamp(vec![1726803356466]),
-                ColumnView::from_bools(vec![true]),
-                ColumnView::from_tiny_ints(vec![1]),
-                ColumnView::from_small_ints(vec![2]),
-                ColumnView::from_ints(vec![3]),
-                ColumnView::from_big_ints(vec![4]),
-                ColumnView::from_floats(vec![5.5]),
-                ColumnView::from_doubles(vec![6.6]),
-                ColumnView::from_unsigned_tiny_ints(vec![7]),
-                ColumnView::from_unsigned_small_ints(vec![8]),
-                ColumnView::from_unsigned_ints(vec![9]),
-                ColumnView::from_unsigned_big_ints(vec![10]),
-                ColumnView::from_varchar(vec!["binary"]),
-                ColumnView::from_nchar(vec!["nchar"]),
-                ColumnView::from_geobytes(vec![vec![
+        let data = Stmt2BindData::new(None, Some(tags), None);
+        let datas = [data];
+
+        let fields = vec![
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+        ];
+
+        let res = bind_datas_as_bytes(&datas, 100, 200, true, Some(&fields), 0)?;
+
+        #[rustfmt::skip]
+        let expected = [
+            // fixed headers
+            0x64, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // req_id
+            0xc8, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // stmt_id
+            0x09, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // action
+            0x01, 0x00, // version
+            0xff, 0xff, 0xff, 0xff, // col_idx
+
+            // data
+            0xac, 0x01, 0x00, 0x00, // TotalLength
+            0x01, 0x00, 0x00, 0x00, // TableCount
+            0x10, 0x00, 0x00, 0x00, // TagCount
+            0x00, 0x00, 0x00, 0x00, // ColCount
+            0x00, 0x00, 0x00, 0x00, // TableNamesOffset
+            0x1c, 0x00, 0x00, 0x00, // TagsOffset
+            0x00, 0x00, 0x00, 0x00, // ColsOffset
+
+            // tags
+            // TagsDataLength
+            0x8c, 0x01, 0x00, 0x00,
+            // TagsBuffer
+            // table 0 tags
+            // tag 0
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x09, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x32, 0x2b, 0x80, 0x0d, 0x92, 0x01, 0x00, 0x00, // Buffer
+
+            // tag 1
+            0x13, 0x00, 0x00, 0x00, // TotalLength
+            0x01, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x01, 0x00, 0x00, 0x00, // BufferLength
+            0x01, // Buffer
+
+            // tag 2
+            0x13, 0x00, 0x00, 0x00, // TotalLength
+            0x02, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x01, 0x00, 0x00, 0x00, // BufferLength
+            0x01, // Buffer
+
+            // tag 3
+            0x14, 0x00, 0x00, 0x00, // TotalLength
+            0x03, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x02, 0x00, 0x00, 0x00, // BufferLength
+            0x02, 0x00, // Buffer
+
+            // tag 4
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x04, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x04, 0x00, 0x00, 0x00, // BufferLength
+            0x03, 0x00, 0x00, 0x00, // Buffer
+
+            // tag 5
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x05, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // Buffer
+
+            // tag 6
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x06, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x04, 0x00, 0x00, 0x00, // BufferLength
+            0x00, 0x00, 0xb0, 0x40, // Buffer
+
+            // tag 7
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x07, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x1a, 0x40, // Buffer
+
+            // tag 8
+            0x13, 0x00, 0x00, 0x00, // TotalLength
+            0x0b, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x01, 0x00, 0x00, 0x00, // BufferLength
+            0x07, // Buffer
+
+            // tag 9
+            0x14, 0x00, 0x00, 0x00, // TotalLength
+            0x0c, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x02, 0x00, 0x00, 0x00, // BufferLength
+            0x08, 0x00, // Buffer
+
+            // tag 10
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x0d, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x04, 0x00, 0x00, 0x00, // BufferLength
+            0x09, 0x00, 0x00, 0x00, // Buffer
+
+            // tag 11
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x0e, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x0a, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // Buffer
+
+            // tag 12
+            0x1c, 0x00, 0x00, 0x00, // TotalLength
+            0x08, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            0x06, 0x00, 0x00, 0x00, // Length
+            0x06, 0x00, 0x00, 0x00, // BufferLength
+            0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, // Buffer
+
+            // tag 13
+            0x1b, 0x00, 0x00, 0x00, // TotalLength
+            0x0a, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            0x05, 0x00, 0x00, 0x00, // Length
+            0x05, 0x00, 0x00, 0x00, // BufferLength
+            0x6e, 0x63, 0x68, 0x61, 0x72, // Buffer
+
+            // tag 14
+            0x2b, 0x00, 0x00, 0x00, // TotalLength
+            0x14, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            0x15, 0x00, 0x00, 0x00, // Length
+            0x15, 0x00, 0x00, 0x00, // BufferLength
+            0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, // Buffer
+
+            // tag 15
+            0x1f, 0x00, 0x00, 0x00, // TotalLength
+            0x10, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            0x09, 0x00, 0x00, 0x00, // Length
+            0x09, 0x00, 0x00, 0x00, // BufferLength
+            0x76, 0x61, 0x72, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, // Buffer
+        ];
+
+        assert_eq!(res, expected);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_bind_datas_as_bytes_with_tags_2() -> anyhow::Result<()> {
+        let tags = &[
+            Value::Null(Ty::Timestamp),
+            Value::Null(Ty::Bool),
+            Value::Null(Ty::TinyInt),
+            Value::Null(Ty::SmallInt),
+            Value::Null(Ty::Int),
+            Value::Null(Ty::BigInt),
+            Value::Null(Ty::Float),
+            Value::Null(Ty::Double),
+            Value::Null(Ty::UTinyInt),
+            Value::Null(Ty::USmallInt),
+            Value::Null(Ty::UInt),
+            Value::Null(Ty::UBigInt),
+            Value::Null(Ty::VarChar),
+            Value::Null(Ty::NChar),
+            Value::Null(Ty::Geometry),
+            Value::Null(Ty::VarBinary),
+        ];
+
+        let data = Stmt2BindData::new(None, Some(tags), None);
+        let datas = [data];
+
+        let fields = vec![
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+        ];
+
+        let res = bind_datas_as_bytes(&datas, 100, 200, true, Some(&fields), 0)?;
+
+        #[rustfmt::skip]
+        let expected = [
+            // fixed headers
+            0x64, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // req_id
+            0xc8, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // stmt_id
+            0x09, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // action
+            0x01, 0x00, // version
+            0xff, 0xff, 0xff, 0xff, // col_idx
+
+            // data
+            0x50, 0x01, 0x00, 0x00, // TotalLength
+            0x01, 0x00, 0x00, 0x00, // TableCount
+            0x10, 0x00, 0x00, 0x00, // TagCount
+            0x00, 0x00, 0x00, 0x00, // ColCount
+            0x00, 0x00, 0x00, 0x00, // TableNamesOffset
+            0x1c, 0x00, 0x00, 0x00, // TagsOffset
+            0x00, 0x00, 0x00, 0x00, // ColsOffset
+
+            // tags
+            // TagsDataLength
+            0x30, 0x01, 0x00, 0x00,
+            // TagsBuffer
+            // table 0 tags
+            // tag 0
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x09, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 1
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x01, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 2
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x02, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 3
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x03, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 4
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x04, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 5
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x05, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 6
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x06, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 7
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x07, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 8
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x0b, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 9
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x0c, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 10
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x0d, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 11
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x0e, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 12
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x08, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x01, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // Length
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 13
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x0a, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x01, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // Length
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 14
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x14, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x01, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // Length
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 15
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x10, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x01, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // Length
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+        ];
+
+        assert_eq!(res, expected);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_bind_datas_as_bytes_with_tags_3() -> anyhow::Result<()> {
+        let tags1 = &[
+            Value::Timestamp(Timestamp::Milliseconds(1726803356466)),
+            Value::Bool(true),
+            Value::TinyInt(1),
+            Value::SmallInt(2),
+            Value::Int(3),
+            Value::BigInt(4),
+            Value::Float(5.5),
+            Value::Double(6.6),
+            Value::UTinyInt(7),
+            Value::USmallInt(8),
+            Value::UInt(9),
+            Value::UBigInt(10),
+            Value::VarChar("binary".to_string()),
+            Value::NChar("nchar".to_string()),
+            Value::Geometry(Bytes::from(vec![
+                0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
+            ])),
+            Value::VarBinary(Bytes::from("varbinary".as_bytes())),
+        ];
+
+        let data1 = Stmt2BindData::new(Some("test1"), Some(tags1), None);
+
+        let tags2 = &[
+            Value::Null(Ty::Timestamp),
+            Value::Null(Ty::Bool),
+            Value::Null(Ty::TinyInt),
+            Value::Null(Ty::SmallInt),
+            Value::Null(Ty::Int),
+            Value::Null(Ty::BigInt),
+            Value::Null(Ty::Float),
+            Value::Null(Ty::Double),
+            Value::Null(Ty::UTinyInt),
+            Value::Null(Ty::USmallInt),
+            Value::Null(Ty::UInt),
+            Value::Null(Ty::UBigInt),
+            Value::Null(Ty::VarChar),
+            Value::Null(Ty::NChar),
+            Value::Null(Ty::Geometry),
+            Value::Null(Ty::VarBinary),
+        ];
+
+        let data2 = Stmt2BindData::new(Some("testnil"), Some(tags2), None);
+
+        let tags3 = &[
+            Value::Timestamp(Timestamp::Milliseconds(1726803356466)),
+            Value::Bool(true),
+            Value::TinyInt(1),
+            Value::SmallInt(2),
+            Value::Int(3),
+            Value::BigInt(4),
+            Value::Float(5.5),
+            Value::Double(6.6),
+            Value::UTinyInt(7),
+            Value::USmallInt(8),
+            Value::UInt(9),
+            Value::UBigInt(10),
+            Value::VarChar("binary".to_string()),
+            Value::NChar("nchar".to_string()),
+            Value::Geometry(Bytes::from(vec![
+                0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
+            ])),
+            Value::VarBinary(Bytes::from("varbinary".as_bytes())),
+        ];
+
+        let data3 = Stmt2BindData::new(Some("test2"), Some(tags3), None);
+
+        let datas = [data1, data2, data3];
+
+        let fields = vec![
+            PrepareField {
+                name: "".to_string(),
+                field_type: 1,
+                precision: 0,
+                scale: 0,
+                bytes: 131584,
+                bind_type: BindType::TableName,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+        ];
+
+        let res = bind_datas_as_bytes(&datas, 100, 200, true, Some(&fields), 0)?;
+
+        #[rustfmt::skip]
+        let expected = [
+            // fixed headers
+            0x64, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // req_id
+            0xc8, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // stmt_id
+            0x09, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // action
+            0x01, 0x00, // version
+            0xff, 0xff, 0xff, 0xff, // col_idx
+
+            // data
+            0x8a, 0x04, 0x00, 0x00, // TotalLength
+            0x03, 0x00, 0x00, 0x00, // TableCount
+            0x10, 0x00, 0x00, 0x00, // TagCount
+            0x00, 0x00, 0x00, 0x00, // ColCount
+            0x1c, 0x00, 0x00, 0x00, // TableNamesOffset
+            0x36, 0x00, 0x00, 0x00, // TagsOffset
+            0x00, 0x00, 0x00, 0x00, // ColsOffset
+
+            // table names
+            // TableNameLength
+            0x06, 0x00,
+            0x08, 0x00,
+            0x06, 0x00,
+            // TableNameBuffer
+            0x74, 0x65, 0x73, 0x74, 0x31, 0x00,
+            0x74, 0x65, 0x73, 0x74, 0x6e, 0x69, 0x6c, 0x00,
+            0x74, 0x65, 0x73, 0x74, 0x32, 0x00,
+
+            // tags
+            // TagsDataLength
+            0x8c, 0x01, 0x00, 0x00,
+            0x30, 0x01, 0x00, 0x00,
+            0x8c, 0x01, 0x00, 0x00,
+            // TagsBuffer
+            // table 0 tags
+            // tag 0
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x09, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x32, 0x2b, 0x80, 0x0d, 0x92, 0x01, 0x00, 0x00, // Buffer
+
+            // tag 1
+            0x13, 0x00, 0x00, 0x00, // TotalLength
+            0x01, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x01, 0x00, 0x00, 0x00, // BufferLength
+            0x01, // Buffer
+
+            // tag 2
+            0x13, 0x00, 0x00, 0x00, // TotalLength
+            0x02, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x01, 0x00, 0x00, 0x00, // BufferLength
+            0x01, // Buffer
+
+            // tag 3
+            0x14, 0x00, 0x00, 0x00, // TotalLength
+            0x03, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x02, 0x00, 0x00, 0x00, // BufferLength
+            0x02, 0x00, // Buffer
+
+            // tag 4
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x04, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x04, 0x00, 0x00, 0x00, // BufferLength
+            0x03, 0x00, 0x00, 0x00, // Buffer
+
+            // tag 5
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x05, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // Buffer
+
+            // tag 6
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x06, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x04, 0x00, 0x00, 0x00, // BufferLength
+            0x00, 0x00, 0xb0, 0x40, // Buffer
+
+            // tag 7
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x07, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x1a, 0x40, // Buffer
+
+            // tag 8
+            0x13, 0x00, 0x00, 0x00, // TotalLength
+            0x0b, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x01, 0x00, 0x00, 0x00, // BufferLength
+            0x07, // Buffer
+
+            // tag 9
+            0x14, 0x00, 0x00, 0x00, // TotalLength
+            0x0c, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x02, 0x00, 0x00, 0x00, // BufferLength
+            0x08, 0x00, // Buffer
+
+            // tag 10
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x0d, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x04, 0x00, 0x00, 0x00, // BufferLength
+            0x09, 0x00, 0x00, 0x00, // Buffer
+
+            // tag 11
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x0e, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x0a, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // Buffer
+
+            // tag 12
+            0x1c, 0x00, 0x00, 0x00, // TotalLength
+            0x08, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            0x06, 0x00, 0x00, 0x00, // Length
+            0x06, 0x00, 0x00, 0x00, // BufferLength
+            0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, // Buffer
+
+            // tag 13
+            0x1b, 0x00, 0x00, 0x00, // TotalLength
+            0x0a, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            0x05, 0x00, 0x00, 0x00, // Length
+            0x05, 0x00, 0x00, 0x00, // BufferLength
+            0x6e, 0x63, 0x68, 0x61, 0x72, // Buffer
+
+            // tag 14
+            0x2b, 0x00, 0x00, 0x00, // TotalLength
+            0x14, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            0x15, 0x00, 0x00, 0x00, // Length
+            0x15, 0x00, 0x00, 0x00, // BufferLength
+            0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, // Buffer
+
+            // tag 15
+            0x1f, 0x00, 0x00, 0x00, // TotalLength
+            0x10, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            0x09, 0x00, 0x00, 0x00, // Length
+            0x09, 0x00, 0x00, 0x00, // BufferLength
+            0x76, 0x61, 0x72, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, // Buffer
+
+            // table 1 tags
+            // tag 0
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x09, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 1
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x01, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 2
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x02, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 3
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x03, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 4
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x04, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 5
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x05, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 6
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x06, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 7
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x07, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 8
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x0b, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 9
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x0c, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 10
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x0d, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 11
+            0x12, 0x00, 0x00, 0x00, // TotalLength
+            0x0e, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x00, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 12
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x08, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x01, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // Length
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 13
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x0a, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x01, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // Length
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 14
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x14, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x01, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // Length
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // tag 15
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x10, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x01, // IsNull
+            0x01, // HaveLength
+            0x00, 0x00, 0x00, 0x00, // Length
+            0x00, 0x00, 0x00, 0x00, // BufferLength
+
+            // table 2 tags
+            // tag 0
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x09, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x32, 0x2b, 0x80, 0x0d, 0x92, 0x01, 0x00, 0x00, // Buffer
+
+            // tag 1
+            0x13, 0x00, 0x00, 0x00, // TotalLength
+            0x01, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x01, 0x00, 0x00, 0x00, // BufferLength
+            0x01, // Buffer
+
+            // tag 2
+            0x13, 0x00, 0x00, 0x00, // TotalLength
+            0x02, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x01, 0x00, 0x00, 0x00, // BufferLength
+            0x01, // Buffer
+
+            // tag 3
+            0x14, 0x00, 0x00, 0x00, // TotalLength
+            0x03, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x02, 0x00, 0x00, 0x00, // BufferLength
+            0x02, 0x00, // Buffer
+
+            // tag 4
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x04, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x04, 0x00, 0x00, 0x00, // BufferLength
+            0x03, 0x00, 0x00, 0x00, // Buffer
+
+            // tag 5
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x05, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // Buffer
+
+            // tag 6
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x06, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x04, 0x00, 0x00, 0x00, // BufferLength
+            0x00, 0x00, 0xb0, 0x40, // Buffer
+
+            // tag 7
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x07, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x1a, 0x40, // Buffer
+
+            // tag 8
+            0x13, 0x00, 0x00, 0x00, // TotalLength
+            0x0b, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x01, 0x00, 0x00, 0x00, // BufferLength
+            0x07, // Buffer
+
+            // tag 9
+            0x14, 0x00, 0x00, 0x00, // TotalLength
+            0x0c, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x02, 0x00, 0x00, 0x00, // BufferLength
+            0x08, 0x00, // Buffer
+
+            // tag 10
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x0d, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x04, 0x00, 0x00, 0x00, // BufferLength
+            0x09, 0x00, 0x00, 0x00, // Buffer
+
+            // tag 11
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x0e, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x0a, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // Buffer
+
+            // tag 12
+            0x1c, 0x00, 0x00, 0x00, // TotalLength
+            0x08, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            0x06, 0x00, 0x00, 0x00, // Length
+            0x06, 0x00, 0x00, 0x00, // BufferLength
+            0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, // Buffer
+
+            // tag 13
+            0x1b, 0x00, 0x00, 0x00, // TotalLength
+            0x0a, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            0x05, 0x00, 0x00, 0x00, // Length
+            0x05, 0x00, 0x00, 0x00, // BufferLength
+            0x6e, 0x63, 0x68, 0x61, 0x72, // Buffer
+
+            // tag 14
+            0x2b, 0x00, 0x00, 0x00, // TotalLength
+            0x14, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            0x15, 0x00, 0x00, 0x00, // Length
+            0x15, 0x00, 0x00, 0x00, // BufferLength
+            0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, // Buffer
+
+            // tag 15
+            0x1f, 0x00, 0x00, 0x00, // TotalLength
+            0x10, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            0x09, 0x00, 0x00, 0x00, // Length
+            0x09, 0x00, 0x00, 0x00, // BufferLength
+            0x76, 0x61, 0x72, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, // Buffer
+        ];
+
+        assert_eq!(res, expected);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_bind_datas_as_bytes_with_cols_1() -> anyhow::Result<()> {
+        let cols = &[
+            ColumnView::from_millis_timestamp(vec![1726803356466]),
+            ColumnView::from_bools(vec![true]),
+            ColumnView::from_tiny_ints(vec![1]),
+            ColumnView::from_small_ints(vec![2]),
+            ColumnView::from_ints(vec![3]),
+            ColumnView::from_big_ints(vec![4]),
+            ColumnView::from_floats(vec![5.5]),
+            ColumnView::from_doubles(vec![6.6]),
+            ColumnView::from_unsigned_tiny_ints(vec![7]),
+            ColumnView::from_unsigned_small_ints(vec![8]),
+            ColumnView::from_unsigned_ints(vec![9]),
+            ColumnView::from_unsigned_big_ints(vec![10]),
+            ColumnView::from_varchar(vec!["binary"]),
+            ColumnView::from_nchar(vec!["nchar"]),
+            ColumnView::from_geobytes(vec![vec![
+                0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
+            ]]),
+            ColumnView::from_bytes(vec!["varbinary".as_bytes()]),
+        ];
+
+        let data = Stmt2BindData::new(None, None, Some(cols));
+
+        let res = bind_datas_as_bytes(&[data], 100, 200, false, None, 16)?;
+
+        #[rustfmt::skip]
+        let expected = [
+            // fixed headers
+            0x64, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // req_id
+            0xc8, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // stmt_id
+            0x09, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // action
+            0x01, 0x00, // version
+            0xff, 0xff, 0xff, 0xff, // col_idx
+
+            // data
+            0xac, 0x01, 0x00, 0x00, // TotalLength
+            0x01, 0x00, 0x00, 0x00, // TableCount
+            0x00, 0x00, 0x00, 0x00, // TagCount
+            0x10, 0x00, 0x00, 0x00, // ColCount
+            0x00, 0x00, 0x00, 0x00, // TableNamesOffset
+            0x00, 0x00, 0x00, 0x00, // TagsOffset
+            0x1c, 0x00, 0x00, 0x00, // ColsOffset
+
+            // cols
+            // ColDataLength
+            0x8c, 0x01, 0x00, 0x00,
+            // ColBuffer
+            // table 0 cols
+            // col 0
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x09, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x32, 0x2b, 0x80, 0x0d, 0x92, 0x01, 0x00, 0x00, // Buffer
+
+            // col 1
+            0x13, 0x00, 0x00, 0x00, // TotalLength
+            0x01, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x01, 0x00, 0x00, 0x00, // BufferLength
+            0x01, // Buffer
+
+            // col 2
+            0x13, 0x00, 0x00, 0x00, // TotalLength
+            0x02, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x01, 0x00, 0x00, 0x00, // BufferLength
+            0x01, // Buffer
+
+            // col 3
+            0x14, 0x00, 0x00, 0x00, // TotalLength
+            0x03, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x02, 0x00, 0x00, 0x00, // BufferLength
+            0x02, 0x00, // Buffer
+
+            // col 4
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x04, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x04, 0x00, 0x00, 0x00, // BufferLength
+            0x03, 0x00, 0x00, 0x00, // Buffer
+
+            // col 5
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x05, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // Buffer
+
+            // col 6
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x06, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x04, 0x00, 0x00, 0x00, // BufferLength
+            0x00, 0x00, 0xb0, 0x40, // Buffer
+
+            // col 7
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x07, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x1a, 0x40, // Buffer
+
+            // col 8
+            0x13, 0x00, 0x00, 0x00, // TotalLength
+            0x0b, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x01, 0x00, 0x00, 0x00, // BufferLength
+            0x07, // Buffer
+
+            // col 9
+            0x14, 0x00, 0x00, 0x00, // TotalLength
+            0x0c, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x02, 0x00, 0x00, 0x00, // BufferLength
+            0x08, 0x00, // Buffer
+
+            // col 10
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x0d, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x04, 0x00, 0x00, 0x00, // BufferLength
+            0x09, 0x00, 0x00, 0x00, // Buffer
+
+            // col 11
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x0e, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x0a, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // Buffer
+
+            // col 12
+            0x1c, 0x00, 0x00, 0x00, // TotalLength
+            0x08, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            // Length
+            0x06, 0x00, 0x00, 0x00,
+            0x06, 0x00, 0x00, 0x00, // BufferLength
+            0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, // Buffer
+
+            // col 13
+            0x1b, 0x00, 0x00, 0x00, // TotalLength
+            0x0a, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            // Length
+            0x05, 0x00, 0x00, 0x00,
+            0x05, 0x00, 0x00, 0x00, // BufferLength
+            0x6e, 0x63, 0x68, 0x61, 0x72, // Buffer
+
+            // col 14
+            0x2b, 0x00, 0x00, 0x00, // TotalLength
+            0x14, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            // Length
+            0x15, 0x00, 0x00, 0x00,
+            0x15, 0x00, 0x00, 0x00, // BufferLength
+            0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, // Buffer
+
+            // col 15
+            0x1f, 0x00, 0x00, 0x00, // TotalLength
+            0x10, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            // Length
+            0x09, 0x00, 0x00, 0x00,
+            0x09, 0x00, 0x00, 0x00, // BufferLength
+            0x76, 0x61, 0x72, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, // Buffer
+        ];
+
+        assert_eq!(res, expected);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_bind_datas_as_bytes_with_cols_2() -> anyhow::Result<()> {
+        let tags = &[
+            Value::Timestamp(Timestamp::Milliseconds(1726803356466)),
+            Value::Bool(true),
+            Value::TinyInt(1),
+            Value::SmallInt(2),
+            Value::Int(3),
+            Value::BigInt(4),
+            Value::Float(5.5),
+            Value::Double(6.6),
+            Value::UTinyInt(7),
+            Value::USmallInt(8),
+            Value::UInt(9),
+            Value::UBigInt(10),
+            Value::VarChar("binary".to_string()),
+            Value::NChar("nchar".to_string()),
+            Value::Geometry(Bytes::from(vec![
+                0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
+            ])),
+            Value::VarBinary(Bytes::from("varbinary".as_bytes())),
+        ];
+
+        let cols = &[
+            ColumnView::from_millis_timestamp(vec![1726803356466, 1726803357466, 1726803358466]),
+            ColumnView::from_bools(vec![Some(true), None, Some(false)]),
+            ColumnView::from_tiny_ints(vec![Some(11), None, Some(12)]),
+            ColumnView::from_small_ints(vec![Some(11), None, Some(12)]),
+            ColumnView::from_ints(vec![Some(11), None, Some(12)]),
+            ColumnView::from_big_ints(vec![Some(11), None, Some(12)]),
+            ColumnView::from_floats(vec![Some(11.2), None, Some(12.2)]),
+            ColumnView::from_doubles(vec![Some(11.2), None, Some(12.2)]),
+            ColumnView::from_unsigned_tiny_ints(vec![Some(11), None, Some(12)]),
+            ColumnView::from_unsigned_small_ints(vec![Some(11), None, Some(12)]),
+            ColumnView::from_unsigned_ints(vec![Some(11), None, Some(12)]),
+            ColumnView::from_unsigned_big_ints(vec![Some(11), None, Some(12)]),
+            ColumnView::from_varchar::<&str, _, _, _>(vec![Some("binary1"), None, Some("binary2")]),
+            ColumnView::from_nchar::<&str, _, _, _>(vec![Some("nchar1"), None, Some("nchar2")]),
+            ColumnView::from_geobytes::<Vec<u8>, _, _, _>(vec![
+                Some(vec![
                     0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
                     0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
-                ]]),
-                ColumnView::from_bytes(vec!["varbinary".as_bytes()]),
-            ];
-
-            let data1 = Stmt2BindData::new(Some("test1"), Some(tags1), &[]);
-
-            let tags2 = &[
-                ColumnView::from_millis_timestamp(vec![None]),
-                ColumnView::from_bools(vec![None]),
-                ColumnView::from_tiny_ints(vec![None]),
-                ColumnView::from_small_ints(vec![None]),
-                ColumnView::from_ints(vec![None]),
-                ColumnView::from_big_ints(vec![None]),
-                ColumnView::from_floats(vec![None]),
-                ColumnView::from_doubles(vec![None]),
-                ColumnView::from_unsigned_tiny_ints(vec![None]),
-                ColumnView::from_unsigned_small_ints(vec![None]),
-                ColumnView::from_unsigned_ints(vec![None]),
-                ColumnView::from_unsigned_big_ints(vec![None]),
-                ColumnView::from_varchar::<&str, _, _, _>(vec![None]),
-                ColumnView::from_nchar::<&str, _, _, _>(vec![None]),
-                ColumnView::from_geobytes::<Vec<u8>, _, _, _>(vec![None]),
-                ColumnView::from_bytes::<Vec<u8>, _, _, _>(vec![None]),
-            ];
-
-            let data2 = Stmt2BindData::new(Some("testnil"), Some(tags2), &[]);
-
-            let tags3 = &[
-                ColumnView::from_millis_timestamp(vec![1726803356466]),
-                ColumnView::from_bools(vec![true]),
-                ColumnView::from_tiny_ints(vec![1]),
-                ColumnView::from_small_ints(vec![2]),
-                ColumnView::from_ints(vec![3]),
-                ColumnView::from_big_ints(vec![4]),
-                ColumnView::from_floats(vec![5.5]),
-                ColumnView::from_doubles(vec![6.6]),
-                ColumnView::from_unsigned_tiny_ints(vec![7]),
-                ColumnView::from_unsigned_small_ints(vec![8]),
-                ColumnView::from_unsigned_ints(vec![9]),
-                ColumnView::from_unsigned_big_ints(vec![10]),
-                ColumnView::from_varchar(vec!["binary"]),
-                ColumnView::from_nchar(vec!["nchar"]),
-                ColumnView::from_geobytes(vec![vec![
+                ]),
+                None,
+                Some(vec![
                     0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
                     0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
-                ]]),
-                ColumnView::from_bytes(vec!["varbinary".as_bytes()]),
-            ];
-
-            let data3 = Stmt2BindData::new(Some("test2"), Some(tags3), &[]);
-
-            let res = bind_datas_as_bytes(&[data1, data2, data3], true)?;
-
-            #[rustfmt::skip]
-            let expected = [
-                0x8a, 0x04, 0x00, 0x00, // total length
-                0x03, 0x00, 0x00, 0x00, // table count
-                0x10, 0x00, 0x00, 0x00, // tag count
-                0x00, 0x00, 0x00, 0x00, // col count
-                0x1c, 0x00, 0x00, 0x00, // table names offset
-                0x36, 0x00, 0x00, 0x00, // tags offset
-                0x00, 0x00, 0x00, 0x00, // col offset
-
-                // table names
-
-                // table name length
-                0x06, 0x00,
-                0x08, 0x00,
-                0x06, 0x00,
-                0x74, 0x65, 0x73, 0x74, 0x31, 0x00, // test1
-                0x74, 0x65, 0x73, 0x74, 0x6e, 0x69, 0x6c, 0x00, // testnil
-                0x74, 0x65, 0x73, 0x74, 0x32, 0x00, // test2
-
-                // tags
-
-                // tags data length
-                0x8c, 0x01, 0x00, 0x00, // table1 data length
-                0x30, 0x01, 0x00, 0x00, // table2 data length
-                0x8c, 0x01, 0x00, 0x00, // table3 data length
-
-                // tagsdata
-                // table1 tags
-                // tag1 timestamp
-                0x1a, 0x00, 0x00, 0x00, // total length
-                0x09, 0x00, 0x00, 0x00, // type
-                0x01, 0x00, 0x00, 0x00, // num
-                0x00, // isnull
-                0x00, // have length
-                0x08, 0x00, 0x00, 0x00, // buffer length
-                0x32, 0x2b, 0x80, 0x0d, 0x92, 0x01, 0x00, 0x00, // buffer
-
-                // tag2 bool
-                0x13, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-
-                // tag3 tinyint
-                0x13, 0x00, 0x00, 0x00,
-                0x02, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-
-                // tag4 smallint
-                0x14, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x02, 0x00, 0x00, 0x00,
-                0x02, 0x00,
-
-                // tag5 int
-                0x16, 0x00, 0x00, 0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-
-                // tag6 bigint
-                0x1a, 0x00, 0x00, 0x00,
-                0x05, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-
-                // tag7 float
-                0x16, 0x00, 0x00, 0x00,
-                0x06, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0xb0, 0x40,
-
-                // tag8 double
-                0x1a, 0x00, 0x00, 0x00,
-                0x07, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x1a, 0x40,
-
-                // tag9 utinyint
-                0x13, 0x00, 0x00, 0x00,
-                0x0b, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x07,
-
-                // tag10 usmallint
-                0x14, 0x00, 0x00, 0x00,
-                0x0c, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x02, 0x00, 0x00, 0x00,
-                0x08, 0x00,
-
-                // tag11 uint
-                0x16, 0x00, 0x00, 0x00,
-                0x0d, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x09, 0x00, 0x00, 0x00,
-
-                // tag12 ubigint
-                0x1a, 0x00, 0x00, 0x00,
-                0x0e, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x0a, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-
-                // tag13 binary
-                0x1c, 0x00, 0x00, 0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x01,
-                0x06, 0x00, 0x00, 0x00,
-                0x06, 0x00, 0x00, 0x00,
-                0x62, 0x69, 0x6e, 0x61, 0x72, 0x79,
-
-                // tag14 nchar
-                0x1b, 0x00, 0x00, 0x00,
-                0x0a, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x01,
-                0x05, 0x00, 0x00, 0x00,
-                0x05, 0x00, 0x00, 0x00,
-                0x6e, 0x63, 0x68, 0x61, 0x72,
-
-                // tag15 geometry
-                0x2b, 0x00, 0x00, 0x00,
-                0x14, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x01,
-                0x15, 0x00, 0x00, 0x00,
-                0x15, 0x00, 0x00, 0x00,
-                0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
-
-                // tag16 varbinary
-                0x1f, 0x00, 0x00, 0x00,
-                0x10, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x01,
-                0x09, 0x00, 0x00, 0x00,
-                0x09, 0x00, 0x00, 0x00,
-                0x76, 0x61, 0x72, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79,
-
-                // table 2 tags
-                // tag1 timestamp nil
-                0x12, 0x00, 0x00, 0x00, // total length
-                0x09, 0x00, 0x00, 0x00, // type
-                0x01, 0x00, 0x00, 0x00, // num
-                0x01, // isnull
-                0x00, // have length
-                0x00, 0x00, 0x00, 0x00, // buffer length
-
-                // tag2 bool nil
-                0x12, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-                0x00,
-                0x00, 0x00, 0x00, 0x00,
-
-                // tag3 tinyint nil
-                0x12, 0x00, 0x00, 0x00,
-                0x02, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-                0x00,
-                0x00, 0x00, 0x00, 0x00,
-
-                // tag4 smallint nil
-                0x12, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-                0x00,
-                0x00, 0x00, 0x00, 0x00,
-
-                // tag5 int nil
-                0x12, 0x00, 0x00, 0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-                0x00,
-                0x00, 0x00, 0x00, 0x00,
-
-                //  tag6 bigint nil
-                0x12, 0x00, 0x00, 0x00,
-                0x05, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-                0x00,
-                0x00, 0x00, 0x00, 0x00,
-
-                // tag7 float nil
-                0x12, 0x00, 0x00, 0x00,
-                0x06, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-                0x00,
-                0x00, 0x00, 0x00, 0x00,
-
-                // tag8 double nil
-                0x12, 0x00, 0x00, 0x00,
-                0x07, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-                0x00,
-                0x00, 0x00, 0x00, 0x00,
-
-                // tag9 utinyint nil
-                0x12, 0x00, 0x00, 0x00,
-                0x0b, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-                0x00,
-                0x00, 0x00, 0x00, 0x00,
-
-                // tag10 usmallint nil
-                0x12, 0x00, 0x00, 0x00,
-                0x0c, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-                0x00,
-                0x00, 0x00, 0x00, 0x00,
-
-                // tag11 uint nil
-                0x12, 0x00, 0x00, 0x00,
-                0x0d, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-                0x00,
-                0x00, 0x00, 0x00, 0x00,
-
-                // tag12 ubigint nil
-                0x12, 0x00, 0x00, 0x00,
-                0x0e, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-                0x00,
-                0x00, 0x00, 0x00, 0x00,
-
-                // tag13 binary nil
-                0x16, 0x00, 0x00, 0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-                0x01,
-                0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00,
-
-                // tag14 nchar nil
-                0x16, 0x00, 0x00, 0x00,
-                0x0a, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-                0x01,
-                0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00,
-
-                // tag15 geometry nil
-                0x16, 0x00, 0x00, 0x00,
-                0x14, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-                0x01,
-                0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00,
-
-                // tag16 varbinary nil
-                0x16, 0x00, 0x00, 0x00,
-                0x10, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-                0x01,
-                0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00,
-
-                // table 3 tags
-                // tag1 timestamp
-                0x1a, 0x00, 0x00, 0x00,
-                0x09, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x32, 0x2b, 0x80, 0x0d, 0x92, 0x01, 0x00, 0x00,
-
-                // tag2 bool
-                0x13, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-
-                // tag3 tinyint
-                0x13, 0x00, 0x00, 0x00,
-                0x02, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-
-                // tag4 smallint
-                0x14, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x02, 0x00, 0x00, 0x00,
-                0x02, 0x00,
-
-                // tag5 int
-                0x16, 0x00, 0x00, 0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-
-                // tag6 bigint
-                0x1a, 0x00, 0x00, 0x00,
-                0x05, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x04, 0x00, 0x00, 0x00,
-
-                // tag7 float
-                0x00, 0x00, 0x00, 0x00,
-                0x16, 0x00, 0x00, 0x00,
-                0x06, 0x00, 0x00, 0x00,
-                0x01,
-                0x00,
-                0x00, 0x00, 0x00, 0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0xb0, 0x40,
-
-                // tag8 double
-                0x1a, 0x00, 0x00, 0x00,
-                0x07, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x1a, 0x40,
-
-                // tag9 utinyint
-                0x13, 0x00, 0x00, 0x00,
-                0x0b, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x07,
-
-                // tag10 usmallint
-                0x14, 0x00, 0x00, 0x00,
-                0x0c, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x02, 0x00, 0x00, 0x00,
-                0x08, 0x00,
-
-                // tag11 uint
-                0x16, 0x00, 0x00, 0x00,
-                0x0d, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x09, 0x00, 0x00, 0x00,
-
-                // tag12 ubigint
-                0x1a, 0x00, 0x00, 0x00,
-                0x0e, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x0a, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-
-                // tag13 binary
-                0x1c, 0x00, 0x00, 0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x01,
-                0x06, 0x00, 0x00, 0x00,
-                0x06, 0x00, 0x00, 0x00,
-                0x62, 0x69, 0x6e, 0x61, 0x72, 0x79,
-
-                // tag14 nchar
-                0x1b, 0x00, 0x00, 0x00,
-                0x0a, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x01,
-                0x05, 0x00, 0x00, 0x00,
-                0x05, 0x00, 0x00, 0x00,
-                0x6e, 0x63, 0x68, 0x61, 0x72,
-
-                // tag15 geometry
-                0x2b, 0x00, 0x00, 0x00,
-                0x14, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x01,
-                0x15, 0x00, 0x00, 0x00,
-                0x15, 0x00, 0x00, 0x00,
-                0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
-
-                // tag16 varbinary
-                0x1f, 0x00, 0x00, 0x00,
-                0x10, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x01,
-                0x09, 0x00, 0x00, 0x00,
-                0x09, 0x00, 0x00, 0x00,
-                0x76, 0x61, 0x72, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79,
-            ];
-
-            assert_eq!(res, expected);
-        }
-
-        {
-            let tags = &[
-                ColumnView::from_millis_timestamp(vec![1726803356466]),
-                ColumnView::from_bools(vec![true]),
-                ColumnView::from_tiny_ints(vec![1]),
-                ColumnView::from_small_ints(vec![2]),
-                ColumnView::from_ints(vec![3]),
-                ColumnView::from_big_ints(vec![4]),
-                ColumnView::from_floats(vec![5.5]),
-                ColumnView::from_doubles(vec![6.6]),
-                ColumnView::from_unsigned_tiny_ints(vec![7]),
-                ColumnView::from_unsigned_small_ints(vec![8]),
-                ColumnView::from_unsigned_ints(vec![9]),
-                ColumnView::from_unsigned_big_ints(vec![10]),
-                ColumnView::from_varchar(vec!["binary"]),
-                ColumnView::from_nchar(vec!["nchar"]),
-                ColumnView::from_geobytes(vec![vec![
-                    0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
-                    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
-                ]]),
-                ColumnView::from_bytes(vec!["varbinary".as_bytes()]),
-            ];
-
-            let cols = &[
-                ColumnView::from_millis_timestamp(vec![
-                    1726803356466,
-                    1726803357466,
-                    1726803358466,
                 ]),
-                ColumnView::from_bools(vec![Some(true), None, Some(false)]),
-                ColumnView::from_tiny_ints(vec![Some(11), None, Some(12)]),
-                ColumnView::from_small_ints(vec![Some(11), None, Some(12)]),
-                ColumnView::from_ints(vec![Some(11), None, Some(12)]),
-                ColumnView::from_big_ints(vec![Some(11), None, Some(12)]),
-                ColumnView::from_floats(vec![Some(11.2), None, Some(12.2)]),
-                ColumnView::from_doubles(vec![Some(11.2), None, Some(12.2)]),
-                ColumnView::from_unsigned_tiny_ints(vec![Some(11), None, Some(12)]),
-                ColumnView::from_unsigned_small_ints(vec![Some(11), None, Some(12)]),
-                ColumnView::from_unsigned_ints(vec![Some(11), None, Some(12)]),
-                ColumnView::from_unsigned_big_ints(vec![Some(11), None, Some(12)]),
-                ColumnView::from_varchar::<&str, _, _, _>(vec![
-                    Some("binary1"),
-                    None,
-                    Some("binary2"),
-                ]),
-                ColumnView::from_nchar::<&str, _, _, _>(vec![Some("nchar1"), None, Some("nchar2")]),
-                ColumnView::from_geobytes::<Vec<u8>, _, _, _>(vec![
-                    Some(vec![
-                        0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59,
-                        0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
-                    ]),
-                    None,
-                    Some(vec![
-                        0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59,
-                        0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
-                    ]),
-                ]),
-                ColumnView::from_bytes::<&[u8], _, _, _>(vec![
-                    Some("varbinary1".as_bytes()),
-                    None,
-                    Some("varbinary2".as_bytes()),
-                ]),
-            ];
+            ]),
+            ColumnView::from_bytes::<&[u8], _, _, _>(vec![
+                Some("varbinary1".as_bytes()),
+                None,
+                Some("varbinary2".as_bytes()),
+            ]),
+        ];
 
-            let data = Stmt2BindData::new(Some("test1"), Some(tags), cols);
-            let res = bind_datas_as_bytes(&[data], true)?;
+        let data = Stmt2BindData::new(Some("test1"), Some(tags), Some(cols));
 
-            #[rustfmt::skip]
-            let expected = [
-                0x19, 0x04, 0x00, 0x00, // total length
-                0x01, 0x00, 0x00, 0x00, // table count
-                0x10, 0x00, 0x00, 0x00, // tag count
-                0x10, 0x00, 0x00, 0x00, // col count
-                0x1c, 0x00, 0x00, 0x00, // table names offset
-                0x24, 0x00, 0x00, 0x00, // tags offset
-                0xb4, 0x01, 0x00, 0x00, // cols offset
+        let fields = vec![
+            PrepareField {
+                name: "".to_string(),
+                field_type: 1,
+                precision: 0,
+                scale: 0,
+                bytes: 131584,
+                bind_type: BindType::TableName,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "ts".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Tag,
+            },
+            PrepareField {
+                name: "a".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Column,
+            },
+            PrepareField {
+                name: "a".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Column,
+            },
+            PrepareField {
+                name: "a".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Column,
+            },
+            PrepareField {
+                name: "a".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Column,
+            },
+            PrepareField {
+                name: "a".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Column,
+            },
+            PrepareField {
+                name: "a".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Column,
+            },
+            PrepareField {
+                name: "a".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Column,
+            },
+            PrepareField {
+                name: "a".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Column,
+            },
+            PrepareField {
+                name: "a".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Column,
+            },
+            PrepareField {
+                name: "a".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Column,
+            },
+            PrepareField {
+                name: "a".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Column,
+            },
+            PrepareField {
+                name: "a".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Column,
+            },
+            PrepareField {
+                name: "a".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Column,
+            },
+            PrepareField {
+                name: "a".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Column,
+            },
+            PrepareField {
+                name: "a".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Column,
+            },
+            PrepareField {
+                name: "a".to_string(),
+                field_type: 9,
+                precision: 0,
+                scale: 0,
+                bytes: 8,
+                bind_type: BindType::Column,
+            },
+        ];
 
-                // table names
+        let res = bind_datas_as_bytes(&[data], 100, 200, true, Some(&fields), 0)?;
 
-                0x06, 0x00, // table name length
-                0x74, 0x65, 0x73, 0x74, 0x31, 0x00, // table name buffer
+        #[rustfmt::skip]
+        let expected = [
+            // fixed headers
+            0x64, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // req_id
+            0xc8, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // stmt_id
+            0x09, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // action
+            0x01, 0x00, // version
+            0xff, 0xff, 0xff, 0xff, // col_id
 
-                // tags
+            // data
+            0x19, 0x04, 0x00, 0x00, // TotalLength
+            0x01, 0x00, 0x00, 0x00, // TableCount
+            0x10, 0x00, 0x00, 0x00, // TagCount
+            0x10, 0x00, 0x00, 0x00, // ColCount
+            0x1c, 0x00, 0x00, 0x00, // TableNamesOffset
+            0x24, 0x00, 0x00, 0x00, // TagsOffset
+            0xb4, 0x01, 0x00, 0x00, // ColsOffset
 
-                0x8c, 0x01, 0x00, 0x00, // tags data length
+            // table names
+            // TableNameLength
+            0x06, 0x00,
+            // TableNameBuffer
+            0x74, 0x65, 0x73, 0x74, 0x31, 0x00,
 
-                // tags buffer
-                // tag1 timestamp
-                0x1a, 0x00, 0x00, 0x00, // total length
-                0x09, 0x00, 0x00, 0x00, // type
-                0x01, 0x00, 0x00, 0x00, // num
-                0x00, // isnull
-                0x00, // have length
-                0x08, 0x00, 0x00, 0x00, // buffer length
-                0x32, 0x2b, 0x80, 0x0d, 0x92, 0x01, 0x00, 0x00, // buffer
+            // tags
+            // TagsDataLength
+            0x8c, 0x01, 0x00, 0x00,
+            // TagsBuffer
+            // table 0 tags
+            // tag 0
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x09, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x32, 0x2b, 0x80, 0x0d, 0x92, 0x01, 0x00, 0x00, // Buffer
 
-                // tag2 bool
-                0x13, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
+            // tag 1
+            0x13, 0x00, 0x00, 0x00, // TotalLength
+            0x01, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x01, 0x00, 0x00, 0x00, // BufferLength
+            0x01, // Buffer
 
-                // tag3 tinyint
-                0x13, 0x00, 0x00, 0x00,
-                0x02, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
+            // tag 2
+            0x13, 0x00, 0x00, 0x00, // TotalLength
+            0x02, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x01, 0x00, 0x00, 0x00, // BufferLength
+            0x01, // Buffer
 
-                // tag4 smallint
-                0x14, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x02, 0x00, 0x00, 0x00,
-                0x02, 0x00,
+            // tag 3
+            0x14, 0x00, 0x00, 0x00, // TotalLength
+            0x03, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x02, 0x00, 0x00, 0x00, // BufferLength
+            0x02, 0x00, // Buffer
 
-                // tag5 int
-                0x16, 0x00, 0x00, 0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
+            // tag 4
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x04, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x04, 0x00, 0x00, 0x00, // BufferLength
+            0x03, 0x00, 0x00, 0x00, // Buffer
 
-                // tag6 bigint
-                0x1a, 0x00, 0x00, 0x00,
-                0x05, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+            // tag 5
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x05, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // Buffer
 
-                // tag7 float
-                0x16, 0x00, 0x00, 0x00,
-                0x06, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0xb0, 0x40,
+            // tag 6
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x06, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x04, 0x00, 0x00, 0x00, // BufferLength
+            0x00, 0x00, 0xb0, 0x40, // Buffer
 
-                // tag8 double
-                0x1a, 0x00, 0x00, 0x00,
-                0x07, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x1a, 0x40,
+            // tag 7
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x07, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x1a, 0x40, // Buffer
 
-                // tag9 utinyint
-                0x13, 0x00, 0x00, 0x00,
-                0x0b, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x07,
+            // tag 8
+            0x13, 0x00, 0x00, 0x00, // TotalLength
+            0x0b, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x01, 0x00, 0x00, 0x00, // BufferLength
+            0x07, // Buffer
 
-                // tag10 usmallint
-                0x14, 0x00, 0x00, 0x00,
-                0x0c, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x02, 0x00, 0x00, 0x00,
-                0x08, 0x00,
+            // tag 9
+            0x14, 0x00, 0x00, 0x00, // TotalLength
+            0x0c, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x02, 0x00, 0x00, 0x00, // BufferLength
+            0x08, 0x00, // Buffer
 
-                // tag11 uint
-                0x16, 0x00, 0x00, 0x00,
-                0x0d, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x09, 0x00, 0x00, 0x00,
+            // tag 10
+            0x16, 0x00, 0x00, 0x00, // TotalLength
+            0x0d, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x04, 0x00, 0x00, 0x00, // BufferLength
+            0x09, 0x00, 0x00, 0x00, // Buffer
 
-                // tag12 ubigint
-                0x1a, 0x00, 0x00, 0x00,
-                0x0e, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x0a, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+            // tag 11
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x0e, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x00, // HaveLength
+            0x08, 0x00, 0x00, 0x00, // BufferLength
+            0x0a, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // Buffer
 
-                // tag13 binary
-                0x1c, 0x00, 0x00, 0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x01,
-                0x06, 0x00, 0x00, 0x00,
-                0x06, 0x00, 0x00, 0x00,
-                0x62, 0x69, 0x6e, 0x61, 0x72, 0x79,
+            // tag 12
+            0x1c, 0x00, 0x00, 0x00, // TotalLength
+            0x08, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            0x06, 0x00, 0x00, 0x00, // Length
+            0x06, 0x00, 0x00, 0x00, // BufferLength
+            0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, // Buffer
 
-                // tag14 nchar
-                0x1b, 0x00, 0x00, 0x00,
-                0x0a, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x01,
-                0x05, 0x00, 0x00, 0x00,
-                0x05, 0x00, 0x00, 0x00,
-                0x6e, 0x63, 0x68, 0x61, 0x72,
+            // tag 13
+            0x1b, 0x00, 0x00, 0x00, // TotalLength
+            0x0a, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            0x05, 0x00, 0x00, 0x00, // Length
+            0x05, 0x00, 0x00, 0x00, // BufferLength
+            0x6e, 0x63, 0x68, 0x61, 0x72, // Buffer
 
-                // tag15 geometry
-                0x2b, 0x00, 0x00, 0x00,
-                0x14, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x01,
-                0x15, 0x00, 0x00, 0x00,
-                0x15, 0x00, 0x00, 0x00,
-                0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
+            // tag 14
+            0x2b, 0x00, 0x00, 0x00, // TotalLength
+            0x14, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            0x15, 0x00, 0x00, 0x00, // Length
+            0x15, 0x00, 0x00, 0x00, // BufferLength
+            0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, // Buffer
 
-                // tag16 varbinary
-                0x1f, 0x00, 0x00, 0x00,
-                0x10, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x01,
-                0x09, 0x00, 0x00, 0x00,
-                0x09, 0x00, 0x00, 0x00,
-                0x76, 0x61, 0x72, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79,
+            // tag 15
+            0x1f, 0x00, 0x00, 0x00, // TotalLength
+            0x10, 0x00, 0x00, 0x00, // Type
+            0x01, 0x00, 0x00, 0x00, // Num
+            0x00, // IsNull
+            0x01, // HaveLength
+            0x09, 0x00, 0x00, 0x00, // Length
+            0x09, 0x00, 0x00, 0x00, // BufferLength
+            0x76, 0x61, 0x72, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, // Buffer
 
-                // cols
+            // cols
+            // ColDataLength
+            0x61, 0x02, 0x00, 0x00,
+            // ColBuffer
+            // table 0 cols
+            // col 0
+            0x2c, 0x00, 0x00, 0x00, // TotalLength
+            0x09, 0x00, 0x00, 0x00, // Type
+            0x03, 0x00, 0x00, 0x00, // Num
+            0x00, 0x00, 0x00, // IsNull
+            0x00, // HaveLength
+            0x18, 0x00, 0x00, 0x00, // BufferLength
+            0x32, 0x2b, 0x80, 0x0d, 0x92, 0x01, 0x00, 0x00, 0x1a, 0x2f, 0x80, 0x0d, 0x92, 0x01, 0x00, 0x00, 0x02, 0x33, 0x80, 0x0d, 0x92, 0x01, 0x00, 0x00, // Buffer
 
-                0x61, 0x02, 0x00, 0x00, // col data length
+            // col 1
+            0x17, 0x00, 0x00, 0x00, // TotalLength
+            0x01, 0x00, 0x00, 0x00, // Type
+            0x03, 0x00, 0x00, 0x00, // Num
+            0x00, 0x01, 0x00, // IsNull
+            0x00, // HaveLength
+            0x03, 0x00, 0x00, 0x00, // BufferLength
+            0x01, 0x00, 0x00, // Buffer
 
-                // col buffer
-                // col1 timestamp
-                0x2c, 0x00, 0x00, 0x00, // total length
-                0x09, 0x00, 0x00, 0x00, // type
-                0x03, 0x00, 0x00, 0x00, // num
-                0x00, 0x00, 0x00, // isnull
-                0x00, // have length
-                0x18, 0x00, 0x00, 0x00, // buffer length
-                // buffer
-                0x32, 0x2b, 0x80, 0x0d, 0x92, 0x01, 0x00, 0x00,
-                0x1a, 0x2f, 0x80, 0x0d, 0x92, 0x01, 0x00, 0x00,
-                0x02, 0x33, 0x80, 0x0d, 0x92, 0x01, 0x00, 0x00,
+            // col 2
+            0x17, 0x00, 0x00, 0x00, // TotalLength
+            0x02, 0x00, 0x00, 0x00, // Type
+            0x03, 0x00, 0x00, 0x00, // Num
+            0x00, 0x01, 0x00, // IsNull
+            0x00, // HaveLength
+            0x03, 0x00, 0x00, 0x00, // BufferLength
+            0x0b, 0x00, 0x0c, // Buffer
 
-                // col2 bool
-                0x17, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x00, 0x01, 0x00,
-                0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x01,
-                0x00,
-                0x00,
+            // col 3
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x03, 0x00, 0x00, 0x00, // Type
+            0x03, 0x00, 0x00, 0x00, // Num
+            0x00, 0x01, 0x00, // IsNull
+            0x00, // HaveLength
+            0x06, 0x00, 0x00, 0x00, // BufferLength
+            0x0b, 0x00, 0x00, 0x00, 0x0c, 0x00, // Buffer
 
-                // col3 tinyint
-                0x17, 0x00, 0x00, 0x00,
-                0x02, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x00, 0x01, 0x00,
-                0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x0b,
-                0x00,
-                0x0c,
+            // col 4
+            0x20, 0x00, 0x00, 0x00, // TotalLength
+            0x04, 0x00, 0x00, 0x00, // Type
+            0x03, 0x00, 0x00, 0x00, // Num
+            0x00, 0x01, 0x00, // IsNull
+            0x00, // HaveLength
+            0x0c, 0x00, 0x00, 0x00, // BufferLength
+            0x0b, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0c, 0x00, 0x00, 0x00, // Buffer
 
-                // col4 smallint
-                0x1a, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x00, 0x01, 0x00,
-                0x00,
-                0x06, 0x00, 0x00, 0x00,
-                0x0b, 0x00,
-                0x00, 0x00,
-                0x0c, 0x00,
+            // col 5
+            0x2c, 0x00, 0x00, 0x00, // TotalLength
+            0x05, 0x00, 0x00, 0x00, // Type
+            0x03, 0x00, 0x00, 0x00, // Num
+            0x00, 0x01, 0x00, // IsNull
+            0x00, // HaveLength
+            0x18, 0x00, 0x00, 0x00, // BufferLength
+            0x0b, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0c, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // Buffer
 
-                // col5 int
-                0x20, 0x00, 0x00, 0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x00, 0x01, 0x00,
-                0x00,
-                0x0c, 0x00, 0x00, 0x00,
-                0x0b, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00,
-                0x0c, 0x00, 0x00, 0x00,
+            // col 6
+            0x20, 0x00, 0x00, 0x00, // TotalLength
+            0x06, 0x00, 0x00, 0x00, // Type
+            0x03, 0x00, 0x00, 0x00, // Num
+            0x00, 0x01, 0x00, // IsNull
+            0x00, // HaveLength
+            0x0c, 0x00, 0x00, 0x00, // BufferLength
+            0x33, 0x33, 0x33, 0x41, 0x00, 0x00, 0x00, 0x00, 0x33, 0x33, 0x43, 0x41, // Buffer
 
-                // col6 bigint
-                0x2c, 0x00, 0x00, 0x00,
-                0x05, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x00, 0x01, 0x00,
-                0x00,
-                0x18, 0x00, 0x00, 0x00,
-                0x0b, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x0c, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+            // col 7
+            0x2c, 0x00, 0x00, 0x00, // TotalLength
+            0x07, 0x00, 0x00, 0x00, // Type
+            0x03, 0x00, 0x00, 0x00, // Num
+            0x00, 0x01, 0x00, // IsNull
+            0x00, // HaveLength
+            0x18, 0x00, 0x00, 0x00, // BufferLength
+            0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x26, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x28, 0x40, // Buffer
 
-                // col7 float
-                0x20, 0x00, 0x00, 0x00,
-                0x06, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x00, 0x01, 0x00,
-                0x00,
-                0x0c, 0x00, 0x00, 0x00,
-                0x33, 0x33, 0x33, 0x41,
-                0x00, 0x00, 0x00, 0x00,
-                0x33, 0x33, 0x43, 0x41,
+            // col 8
+            0x17, 0x00, 0x00, 0x00, // TotalLength
+            0x0b, 0x00, 0x00, 0x00, // Type
+            0x03, 0x00, 0x00, 0x00, // Num
+            0x00, 0x01, 0x00, // IsNull
+            0x00, // HaveLength
+            0x03, 0x00, 0x00, 0x00, // BufferLength
+            0x0b, 0x00, 0x0c, // Buffer
 
-                // col8 double
-                0x2c, 0x00, 0x00, 0x00,
-                0x07, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x00, 0x01, 0x00,
-                0x00,
-                0x18, 0x00, 0x00, 0x00,
-                0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x26, 0x40,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x28, 0x40,
+            // col 9
+            0x1a, 0x00, 0x00, 0x00, // TotalLength
+            0x0c, 0x00, 0x00, 0x00, // Type
+            0x03, 0x00, 0x00, 0x00, // Num
+            0x00, 0x01, 0x00, // IsNull
+            0x00, // HaveLength
+            0x06, 0x00, 0x00, 0x00, // BufferLength
+            0x0b, 0x00, 0x00, 0x00, 0x0c, 0x00, // Buffer
 
-                // col9 utinyint
-                0x17, 0x00, 0x00, 0x00,
-                0x0b, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x00, 0x01, 0x00,
-                0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x0b,
-                0x00,
-                0x0c,
+            // col 10
+            0x20, 0x00, 0x00, 0x00, // TotalLength
+            0x0d, 0x00, 0x00, 0x00, // Type
+            0x03, 0x00, 0x00, 0x00, // Num
+            0x00, 0x01, 0x00, // IsNull
+            0x00, // HaveLength
+            0x0c, 0x00, 0x00, 0x00, // BufferLength
+            0x0b, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0c, 0x00, 0x00, 0x00, // Buffer
 
-                // col10 usmallint
-                0x1a, 0x00, 0x00, 0x00,
-                0x0c, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x00, 0x01, 0x00,
-                0x00,
-                0x06, 0x00, 0x00, 0x00,
-                0x0b, 0x00,
-                0x00, 0x00,
-                0x0c, 0x00,
+            // col 11
+            0x2c, 0x00, 0x00, 0x00, // TotalLength
+            0x0e, 0x00, 0x00, 0x00, // Type
+            0x03, 0x00, 0x00, 0x00, // Num
+            0x00, 0x01, 0x00, // IsNull
+            0x00, // HaveLength
+            0x18, 0x00, 0x00, 0x00, // BufferLength
+            0x0b, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0c, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // Buffer
 
-                // col11 uint
-                0x20, 0x00, 0x00, 0x00,
-                0x0d, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x00, 0x01, 0x00,
-                0x00,
-                0x0c, 0x00, 0x00, 0x00,
-                0x0b, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00,
-                0x0c, 0x00, 0x00, 0x00,
+            // col 12
+            0x2e, 0x00, 0x00, 0x00, // TotalLength
+            0x08, 0x00, 0x00, 0x00, // Type
+            0x03, 0x00, 0x00, 0x00, // Num
+            0x00, 0x01, 0x00, // IsNull
+            0x01, // HaveLength
+            // Length
+            0x07, 0x00, 0x00, 0x00,
+            0x00, 0x00, 0x00, 0x00,
+            0x07, 0x00, 0x00, 0x00,
+            0x0e, 0x00, 0x00, 0x00, // BufferLength
+            0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x31, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x32, // Buffer
 
-                // col12 ubigint
-                0x2C, 0x00, 0x00, 0x00,
-                0x0e, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x00, 0x01, 0x00,
-                0x00,
-                0x18, 0x00, 0x00, 0x00,
-                0x0b, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x0c, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+            // col 13
+            0x2c, 0x00, 0x00, 0x00, // TotalLength
+            0x0a, 0x00, 0x00, 0x00, // Type
+            0x03, 0x00, 0x00, 0x00, // Num
+            0x00, 0x01, 0x00, // IsNull
+            0x01, // HaveLength
+            // Length
+            0x06, 0x00, 0x00, 0x00,
+            0x00, 0x00, 0x00, 0x00,
+            0x06, 0x00, 0x00, 0x00,
+            0x0c, 0x00, 0x00, 0x00, // BufferLength
+            0x6e, 0x63, 0x68, 0x61, 0x72, 0x31, 0x6e, 0x63, 0x68, 0x61, 0x72, 0x32, // Buffer
 
-                // col13 binary
-                0x2e, 0x00, 0x00, 0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x00, 0x01, 0x00,
-                0x01,
-                0x07, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00,
-                0x07, 0x00, 0x00, 0x00,
-                0x0e, 0x00, 0x00, 0x00,
-                0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x31,
-                0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x32,
+            // col 14
+            0x4a, 0x00, 0x00, 0x00, // TotalLength
+            0x14, 0x00, 0x00, 0x00, // Type
+            0x03, 0x00, 0x00, 0x00, // Num
+            0x00, 0x01, 0x00, // IsNull
+            0x01, // HaveLength
+            // Length
+            0x15, 0x00, 0x00, 0x00,
+            0x00, 0x00, 0x00, 0x00,
+            0x15, 0x00, 0x00, 0x00,
+            0x2a, 0x00, 0x00, 0x00, // BufferLength
+            0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, // Buffer
 
-                // col14 nchar
-                0x2c, 0x00, 0x00, 0x00,
-                0x0a, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x00, 0x01, 0x00,
-                0x01,
-                0x06, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00,
-                0x06, 0x00, 0x00, 0x00,
-                0x0c, 0x00, 0x00, 0x00,
-                0x6e, 0x63, 0x68, 0x61, 0x72, 0x31,
-                0x6e, 0x63, 0x68, 0x61, 0x72, 0x32,
+            // col 15
+            0x34, 0x00, 0x00, 0x00, // TotalLength
+            0x10, 0x00, 0x00, 0x00, // Type
+            0x03, 0x00, 0x00, 0x00, // Num
+            0x00, 0x01, 0x00, // IsNull
+            0x01, // HaveLength
+            // Length
+            0x0a, 0x00, 0x00, 0x00,
+            0x00, 0x00, 0x00, 0x00,
+            0x0a, 0x00, 0x00, 0x00,
+            0x14, 0x00, 0x00, 0x00, // BufferLength
+            0x76, 0x61, 0x72, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x31, 0x76, 0x61, 0x72, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x32, // Buffer
+        ];
 
-                // col15 geometry
-                0x4a, 0x00, 0x00, 0x00,
-                0x14, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x00, 0x01, 0x00,
-                0x01,
-                0x15, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00,
-                0x15, 0x00, 0x00, 0x00,
-                0x2a, 0x00, 0x00, 0x00,
-                0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
-                0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
-
-                // col16 varbinary
-                0x34, 0x00, 0x00, 0x00,
-                0x10, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x00, 0x01, 0x00,
-                0x01,
-                0x0a, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00,
-                0x0a, 0x00, 0x00, 0x00,
-                0x14, 0x00, 0x00, 0x00,
-                0x76, 0x61, 0x72, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x31,
-                0x76, 0x61, 0x72, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79, 0x32,
-            ];
-
-            assert_eq!(res, expected);
-        }
-
-        {
-            let cols = &[
-                ColumnView::from_millis_timestamp(vec![1726803356466]),
-                ColumnView::from_bools(vec![true]),
-                ColumnView::from_tiny_ints(vec![1]),
-                ColumnView::from_small_ints(vec![2]),
-                ColumnView::from_ints(vec![3]),
-                ColumnView::from_big_ints(vec![4]),
-                ColumnView::from_floats(vec![5.5]),
-                ColumnView::from_doubles(vec![6.6]),
-                ColumnView::from_unsigned_tiny_ints(vec![7]),
-                ColumnView::from_unsigned_small_ints(vec![8]),
-                ColumnView::from_unsigned_ints(vec![9]),
-                ColumnView::from_unsigned_big_ints(vec![10]),
-                ColumnView::from_varchar(vec!["binary"]),
-                ColumnView::from_nchar(vec!["nchar"]),
-                ColumnView::from_geobytes(vec![vec![
-                    0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
-                    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
-                ]]),
-                ColumnView::from_bytes(vec!["varbinary".as_bytes()]),
-            ];
-
-            let data = Stmt2BindData::new(None, None, cols);
-            let res = bind_datas_as_bytes(&[data], false)?;
-
-            #[rustfmt::skip]
-            let expected = [
-                0xac, 0x01, 0x00, 0x00, // total length
-                0x01, 0x00, 0x00, 0x00, // table count
-                0x00, 0x00, 0x00, 0x00, // tag count
-                0x10, 0x00, 0x00, 0x00, // col count
-                0x00, 0x00, 0x00, 0x00, // table names offset
-                0x00, 0x00, 0x00, 0x00, // tags offset
-                0x1c, 0x00, 0x00, 0x00, // cols offset
-
-                // cols
-
-                0x8c, 0x01, 0x00, 0x00, // col data length
-                // table 0 cols
-                // col 0
-                0x1a, 0x00, 0x00, 0x00, // total length
-                0x09, 0x00, 0x00, 0x00, // type
-                0x01, 0x00, 0x00, 0x00, // num
-                0x00, // isnull
-                0x00, // have length
-                0x08, 0x00, 0x00, 0x00, // buffer length
-                0x32, 0x2b, 0x80, 0x0d, 0x92, 0x01, 0x00, 0x00, // buffer
-
-                // col 1
-                0x13, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-
-                // col 2
-                0x13, 0x00, 0x00, 0x00,
-                0x02, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x01,
-
-                // col 3
-                0x14, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x02, 0x00, 0x00, 0x00,
-                0x02, 0x00,
-
-                // col 4
-                0x16, 0x00, 0x00, 0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x03, 0x00, 0x00, 0x00,
-
-                // col 5
-                0x1a, 0x00, 0x00, 0x00,
-                0x05, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-
-                // col 6
-                0x16, 0x00, 0x00, 0x00,
-                0x06, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0xb0, 0x40,
-
-                // col 7
-                0x1a, 0x00, 0x00, 0x00,
-                0x07, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x1a, 0x40,
-
-                // col 8
-                0x13, 0x00, 0x00, 0x00,
-                0x0b, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x07,
-
-                // col 9
-                0x14, 0x00, 0x00, 0x00,
-                0x0c, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x02, 0x00, 0x00, 0x00,
-                0x08, 0x00,
-
-                // col 10
-                0x16, 0x00, 0x00, 0x00,
-                0x0d, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x04, 0x00, 0x00, 0x00,
-                0x09, 0x00, 0x00, 0x00,
-
-                // col 11
-                0x1a, 0x00, 0x00, 0x00,
-                0x0e, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x0a, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-
-                // col 12
-                0x1c, 0x00, 0x00, 0x00,
-                0x08, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x01,
-                0x06, 0x00, 0x00, 0x00,
-                0x06, 0x00, 0x00, 0x00,
-                0x62, 0x69, 0x6e, 0x61, 0x72, 0x79,
-
-                // col 13
-                0x1b, 0x00, 0x00, 0x00,
-                0x0a, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x01,
-                0x05, 0x00, 0x00, 0x00,
-                0x05, 0x00, 0x00, 0x00,
-                0x6e, 0x63, 0x68, 0x61, 0x72,
-
-                // col 14
-                0x2b, 0x00, 0x00, 0x00,
-                0x14, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x01,
-                0x15, 0x00, 0x00, 0x00,
-                0x15, 0x00, 0x00, 0x00,
-                0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x59, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40,
-
-                // col 15
-                0x1f, 0x00, 0x00, 0x00,
-                0x10, 0x00, 0x00, 0x00,
-                0x01, 0x00, 0x00, 0x00,
-                0x00,
-                0x01,
-                0x09, 0x00, 0x00, 0x00,
-                0x09, 0x00, 0x00, 0x00,
-                0x76, 0x61, 0x72, 0x62, 0x69, 0x6e, 0x61, 0x72, 0x79,
-            ];
-
-            assert_eq!(res, expected);
-        }
+        assert_eq!(res, expected);
 
         Ok(())
     }
